% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addParamClass.R
\docType{class}
\name{addParamClass}
\alias{addParamClass}
\alias{addParamClass-class}
\title{Class "addParamClass"}
\description{
A Class holding information about the parameters used.
}
\section{Slots}{

\describe{
\item{\code{bin_pseudotime_colname}}{A character string representing the column name for binned pseudotime values.}

\item{\code{path_prefix}}{A character string representing the prefix for path labeling.}

\item{\code{root_label}}{A character string representing the label for the tree root.}

\item{\code{pseudotime_colname}}{A character string representing the column name for pseudotime values.}

\item{\code{bin_method}}{A character string representing the algorithm used for binning.}

\item{\code{path_colname}}{A character string representing the column name for path values.}

\item{\code{bin_colname}}{A character string representing the column name for bin values.}

\item{\code{bin_size_colname}}{A character string representing the column name for bin sizes.}

\item{\code{bin_members_colname}}{A character string representing the column name for bin members.}

\item{\code{annotation_col}}{A character string representing the column name for cell type annotation.}

\item{\code{g}}{Integer. Number of genes taken in the regression fit.}

\item{\code{Q}}{Numeric. Significance Level.}

\item{\code{min.obs}}{Numeric. Minimum value to estimate the model (degree+1) x Groups + 1. (Default = 6).}

\item{\code{MT.adjust}}{A character string specifying the Pvalue correction method used.}

\item{\code{epsilon}}{Numeric. Convergence tolerance.}

\item{\code{step.method}}{A character string specifying the imputed step method for stepwise regression.}

\item{\code{useWeights}}{A logical value specifying whether to use weights during model fitting.}

\item{\code{offset}}{A logical value specifying whether to use offset during model fitting.}

\item{\code{useInverseWeights}}{A logical value specifying whether to take inverse of the weights.}

\item{\code{logOffset}}{A logical value specifying whether to take the logarithm of the offsets during model fitting.}

\item{\code{logWeights}}{A logical value specifying whether to take the logarithm of the weights during model fitting.}

\item{\code{max_it}}{Integer. Maximum number of iterations to fit the model.}
}}

\keyword{classes}
