% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.pseudotime.R
\name{align.pseudotime}
\alias{align.pseudotime}
\title{Align pseudotime from two different paths}
\usage{
align.pseudotime(
  scmpObj,
  pseudotime_col,
  path_col,
  method = "rescale",
  verbose = TRUE
)
}
\arguments{
\item{pseudotime_col}{Name of the column in `cell.metadata` storing
Pseudotime values. It is generated using `colData` from the \pkg{SingleCellExperiment}
package. (Default is "Pseudotime").}

\item{path_col}{Name of the column in `cell.metadata` storing information
for Path. It is generated using `colData` from the \pkg{SingleCellExperiment} package.
(Default is `path_prefix`)}

\item{method}{Currently supported `scales::rescale`.}

\item{verbose}{Print detailed output in the console. (Default is TRUE)}
}
\description{
`align.pseudotime()` is an internal function that aligns the pseudotime from
two different paths. It first identifies the path which is having the longer
range of pseudotime, and then it uses the range of the shorter path to rescale
the range of the longer one. It uses a simple rescaling mechanism from
`scales::rescale`. In future we expect to support strategies like dynamic time
warping from dtw package.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
\keyword{internal}
