% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_pseudotime.R
\name{align_pseudotime}
\alias{align_pseudotime}
\title{Align pseudotime from two different paths}
\usage{
align_pseudotime(
  scmpObj,
  pseudotime_col,
  path_col,
  method = "rescale",
  verbose = TRUE
)
}
\arguments{
\item{pseudotime_col}{Chacarcter string with the column name in `cell.metadata` storing
Pseudotime values. It is generated using `colData` from the \pkg{SingleCellExperiment}
package. (Default is "Pseudotime")}

\item{path_col}{Chacarcter string with the column name in `cell.metadata` storing information
for Path. It is generated using `colData` from the \pkg{SingleCellExperiment} package.
(Default is `path_prefix`)}

\item{method}{Currently only `scales::rescale` is supported. (Default is "rescale")}

\item{verbose}{Print detailed output in the console. (Default is TRUE)}
}
\description{
`align_pseudotime()` is an internal function that aligns the pseudotime from
two different trajectory paths. It first identifies the path which has a longer
pseudotime range, and then it uses the range of the shorter path to rescale
the range of the longer one. It uses a simple rescaling mechanism from
`scales::rescale`. In future we expect to support strategies like dynamic time
warping from dtw package.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
\keyword{internal}
