% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.scmp.R
\name{as.scmp}
\alias{as.scmp}
\title{Convert 'Cell Dataset' or 'SingleCellExperiment' object to scmpClass}
\usage{
as.scmp(
  object,
  from = "cds",
  path_prefix = "Path",
  root_label = "root",
  pseudotime_colname = "Pseudotime",
  align_pseudotime = FALSE,
  path_colname = path_prefix,
  annotation_colname = "cell_type",
  verbose = TRUE,
  interactive = TRUE,
  additional_params = list(labels_exist = FALSE)
)
}
\arguments{
\item{object}{An S4 object of class `cds/CellDataSet` or `SingleCellExperiment`.}

\item{from}{Character string specifying the class of 'object'. Use "cds" for
`cds/CellDataSet` class and "sce" for `SingleCellExperiment` class.}

\item{path_prefix}{Prefix used to annotate the paths. (Default is "Path").}

\item{root_label}{Label used to annotate root cells. (Default is "root").}

\item{pseudotime_colname}{Name of the column in `cell.metadata` storing
Pseudotime values. It is generated using `colData` from the \pkg{SingleCellExperiment}
package. (Default is "Pseudotime").}

\item{align_pseudotime}{Whether to automatically align two different pseudotimes.
See \code{\link{align_pseudotime}} for more details. (Default is FALSE).}

\item{path_colname}{Name of the column in `cell.metadata` storing information
for Path. It is generated using `colData` from the \pkg{SingleCellExperiment} package.
(Default is `path_prefix`)}

\item{annotation_colname}{Column name in cell level metadata containing cell type
annotations. (Default is "cell_type").}

\item{verbose}{Print detailed output in the console. (Default is TRUE)}

\item{interactive}{Whether to use the shiny application to select paths. (Default is TRUE).}

\item{additional_params}{A named list of additional parameters. See details.}
}
\value{
An instance of the 'scmpClass'.
}
\description{
`as.scmp()` converts a cds/CellDataSet object from Monocle3 or a SingleCellExperiment
object to an instance of the scmpClass object.
}
\details{
Additional Details
}
\examples{
# Load Library
# library(scMaSigPro)
# Step-1: Load a dataset for testing
# This dataset is available as part of the package
# It is simulated with splatter
data("splat.sim", package = "scMaSigPro")

# Step-2: Convert to ScMaSigPro Object
# Here, we convert the sce object to an scMaSigPro object
scmp.sce <- as.scmp(
  object = splat.sim, from = "sce",
  align_pseudotime = TRUE,
  verbose = FALSE,
  additional_params = list(
    labels_exist = TRUE,
    existing_pseudotime_colname = "Step",
    existing_path_colname = "Group"
  )
)

}
\seealso{
`colData` from the \pkg{SingleCellExperiment} package, `new_cell_data_set`
function in \pkg{monocle3}
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
