% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown_elements.R
\docType{data}
\name{author_info}
\alias{author_info}
\title{Markdown Elements for Report generation.}
\format{
An object of class \code{character} of length 1.
}
\usage{
author_info
}
\value{
named lists of markdown elements.
}
\description{
Elements are stored in constant named list and are called dynamically.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
\keyword{internal}
