% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.rows.R
\name{average.rows}
\alias{average.rows}
\title{Average Rows by Match and Index}
\usage{
average.rows(x, index, match, r = 0.7)
}
\arguments{
\item{x}{A matrix.}

\item{index}{Index vector indicating how rows must be averaged.}

\item{match}{Match vector for indexing rows.}

\item{r}{The minimal correlation value between rows to compute the average. Default is 0.7.}
}
\value{
A matrix of averaged rows.
}
\description{
Average Rows by Match and Index
}
\details{
The `average.rows` function matches row names of a matrix with a `match` vector, and then
calculates the average of rows based on the `index` vector.


The rows are averaged only if the Pearson correlation coefficient between all rows of each given
index is greater than 'r'. If not, that group of rows is discarded in the result matrix.
}
\examples{
# Create data matrix for row averaging
x <- matrix(rnorm(30), nrow = 6, ncol = 5)
rownames(x) <- paste("ID", c(1, 2, 11, 12, 19, 20), sep = "")
i <- paste("g", rep(c(1:10), each = 2), sep = "") # index vector
m <- paste("ID", c(1:20), sep = "") # match vector
average.rows(x, i, m, r = 0)

}
\author{
Ana Conesa and Maria Jose Nueda
}
