% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{bAlloc,scmp-method}
\alias{bAlloc,scmp-method}
\title{Set or get the Sparse Column Data of an scmp Object}
\usage{
\S4method{bAlloc}{scmp}(object, value = "missing")
}
\arguments{
\item{object}{An object of class `scmp`.}

\item{value}{The new value for the `colData` slot. Optional for getting.}
}
\value{
`colData` when getting, modified `scmp` object when setting.
}
\description{
Set or get the Sparse Column Data of an scmp Object
}
