% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{bAlloc<-,ScMaSigPro-method}
\alias{bAlloc<-,ScMaSigPro-method}
\title{Replacement method for bAlloc}
\usage{
\S4method{bAlloc}{ScMaSigPro}(object) <- value
}
\arguments{
\item{object}{An object of class `ScMaSigPro`.}

\item{value}{The new value for the `colData` slot.}
}
\value{
Modified `ScMaSigPro` object.
}
\description{
Replacement method for bAlloc
}
