% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{bAlloc<-,scmp-method}
\alias{bAlloc<-,scmp-method}
\title{Replacement method for bAlloc}
\usage{
\S4method{bAlloc}{scmp}(object) <- value
}
\arguments{
\item{object}{An object of class `scmp`.}

\item{value}{The new value for the `colData` slot.}
}
\value{
Modified `scmp` object.
}
\description{
Replacement method for bAlloc
}
