% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{bAlloc<-}
\alias{bAlloc<-}
\title{Replacement method for bAlloc.}
\usage{
bAlloc(object) <- value
}
\arguments{
\item{object}{An object of class \code{\link{ScMaSigPro}}.}

\item{value}{The value of the branch Assignment Matrix.}
}
\value{
Modified `ScMaSigPro` object with new Assignment Matrix.
}
\description{
Replacement method for bAlloc.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
