% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{bAlloc}
\alias{bAlloc}
\title{Get or set the branch allocation matrix}
\usage{
bAlloc(object, value = "missing")
}
\arguments{
\item{object}{An object of class `ScMaSigPro`.}

\item{value}{The new value for the `colData` slot to set. Optional for getting.}
}
\value{
`colData` when getting, modified `ScMaSigPro` object when setting.
}
\description{
Get or set the branch allocation matrix
}
