% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_helpers.R
\name{calc_bin_size}
\alias{calc_bin_size}
\title{Calculate Bin Size Function}
\usage{
calc_bin_size(x, clus_mem_col = "scmp_cluster_members")
}
\arguments{
\item{x}{A data frame containing the "cluster.members" column.}
}
\value{
A numeric value representing the size of the bin (number of elements
in the "cluster.members" column).
}
\description{
`calc_bin_size()` calculates the bin size based on the number of elements in
the "cluster.members" column of the input data frame.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
\keyword{internal}
