% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{cDense<-,scmp-method}
\alias{cDense<-,scmp-method}
\title{Replacement method for cSparse}
\usage{
\S4method{cDense}{scmp}(object) <- value
}
\arguments{
\item{object}{An object of class `scmp`.}

\item{value}{The new value for the `colData` slot.}
}
\value{
Modified `scmp` object.
}
\description{
Replacement method for cSparse
}
