% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{cDense}
\alias{cDense}
\title{Get the data for Dense Slot.}
\usage{
cDense(object, value = "missing")
}
\arguments{
\item{object}{An object of class \code{\link{ScMaSigPro}}.}

\item{value}{The new value for the `colData` slot of Dense. (When Setting)}
}
\value{
`colData` for the `Dense` Slot.
}
\description{
Get or set cell level metadata for the Dense slot of an
\code{\link{ScMaSigPro}} object.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
