% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_string}
\alias{clean_string}
\title{Clean or Check Backticks in a String}
\usage{
clean_string(input_string, action = c("check", "remove"))
}
\arguments{
\item{input_string}{A character string to be processed.}

\item{action}{A character string specifying the action to perform. Either "check" to check for backticks or "remove" to remove backticks. Defaults to "check".}
}
\value{
If action is "check", returns a logical value indicating the presence of backticks. If action is "remove", returns the input string with all backticks removed.
}
\description{
This function either checks for the presence of backticks in a string or removes them, based on the specified action.
}
\keyword{internal}
