% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_pseudotime_bin.R
\name{combine_pseudotime_bin}
\alias{combine_pseudotime_bin}
\title{Combine pseudotime and bin data}
\usage{
combine_pseudotime_bin(path.frame, bin_table)
}
\arguments{
\item{path.frame}{A data frame containing pseudotime data. It must include a 'Step' column.}

\item{bin_table}{A data frame containing bin data. It must include 'from', 'to', 'bin_size', and 'binnedTime' columns.}
}
\value{
A data frame which is a combined version of path.frame and bin_table.
}
\description{
This function takes a pseudotime data frame (path.frame) and a bin data frame (bin_table) and combines them.
It subsets path.frame based on the bin_table 'from' and 'to' values and adds this bin_table information to path.frame. 
The function is parallelized to speed up the process.
}
\examples{
\dontrun{
# Assuming 'path.frame' and 'bin_table' are available and have the correct format
combined_data <- combine_pseudotime_bin(path.frame, bin_table)
}

}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
\keyword{internal}
