% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_path.R
\name{convert_to_path}
\alias{convert_to_path}
\title{Convert Vector Elements to Path Names (Internal)}
\usage{
convert_to_path(vec, path_prefix, root_label)
}
\arguments{
\item{vec}{A character vector where elements may be repeated and
might contain the value "root".}

\item{path_prefix}{Prefix used to annoate the paths, default is "Path".}
}
\value{
A character vector with the same length as the input where
unique elements, excluding "root", are renamed to "Path1", "Path2", etc.
}
\description{
`convert_to_path()` transforms a vector by renaming its unique elements
(excluding the "root" element) to a sequence named "Path1", "Path2", etc.
}
\keyword{internal}
