% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_na.R
\name{count.na}
\alias{count.na}
\title{Count NA Function}
\usage{
count.na(x)
}
\description{
Description:
This function calculates the number of non-missing (non-NA) elements in a vector.
}
\details{
Usage:
count.na(x)

Arguments:
  \describe{
    \item{x}{A vector for which you want to count non-missing (non-NA) elements.}
  }

Details:
The function calculates the number of non-missing elements in the input vector
by subtracting the count of missing (NA) elements from the total length of the vector.

Value:
The function returns a numeric value representing the number of non-missing elements
in the input vector.
}
\examples{
# Sample vector
x <- c(1, 2, NA, 4, NA, 6, 7, NA)

# Count non-missing elements in the vector
result <- count.na(x)
print(result)

}
\keyword{internal}
