% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scmp.R
\name{create_scmp}
\alias{create_scmp}
\title{Create ScMaSigPro Object}
\usage{
create_scmp(
  counts,
  cell_data,
  feature_data,
  bin_counts = NULL,
  bin_cell_data = NULL,
  ptime_col,
  path_col,
  use_as_bin = FALSE
)
}
\arguments{
\item{counts}{A matrix containing the raw expression counts.}

\item{cell_data}{A data frame containing the cell level metadata.}

\item{feature_data}{A data frame containing the feature level metadata.}

\item{bin_counts}{A matrix containing the binned counts. (Only if
use_as_bin == TRUE)}

\item{bin_cell_data}{A data frame containing the binned cell level metadata.
(Only if use_as_bin == TRUE)}

\item{ptime_col}{A character string representing the column name
for inferred Pseudotime values. (Default is "Pseudotime")}

\item{path_col}{A character string representing the column name for branching
path assignment. (Default is `path_prefix`)}

\item{use_as_bin}{A logical indicating to run MaSigPro analysis without
pseudotime binning. Defaults to FALSE.}
}
\value{
An object of class \code{\link{ScMaSigPro}}, with aligned pseudotime.
}
\description{
`create_scmp()` initializes a ScMaSigPro object with the given counts,
cell level metadata, and other optional parameters.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
