% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scmpObj.R
\name{create_scmpObj}
\alias{create_scmpObj}
\title{Create scMaSigProClass Object}
\usage{
create_scmpObj(
  counts,
  cell_data,
  feature_data,
  bin_counts,
  bin_cell_data,
  pseudotime_colname,
  path_colname,
  use_as_bin = FALSE
)
}
\arguments{
\item{counts}{A matrix containing the expression counts.}

\item{cell_data}{A data frame containing the cell metadata.}

\item{feature_data}{A data frame containing the feature metadata. (Currently unused)}

\item{bin_counts}{A matrix containing the binned counts. (Currently unused)}

\item{bin_cell_data}{A data frame containing the binned cell data. (Currently unused)}

\item{pseudotime_colname}{A character string specifying the column name for pseudotime in the cell metadata.}

\item{path_colname}{A character string specifying the column name for the path in the cell metadata.}

\item{use_as_bin}{A logical flag indicating whether to use the counts and cell data as bin. Defaults to FALSE.}
}
\value{
A scMaSigProClass object containing the inputted counts, cell data, and additional parameters.
}
\description{
This function initializes a scMaSigProClass object with the given counts, cell data, and other optional parameters.
It is designed for single-cell RNA-seq data analysis.
}
\examples{
\dontrun{
# Assuming you have 'counts_matrix', 'cell_data_df', 'pseudotime_col', and 'path_col'
scmp_object <- create_scmp(counts_matrix, cell_data_df,
  feature_data = NULL,
  bin_counts = NULL, bin_cell_data = NULL,
  pseudotime_col, path_col, use_as_bin = FALSE
)
}

}
