% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{cSparse,scmp-method}
\alias{cSparse,scmp-method}
\title{Set the Sparse Column Data of an scmp Object}
\usage{
\S4method{cSparse}{scmp}(object, value = "missing")
}
\arguments{
\item{object}{An object of class `scmp`.}

\item{value}{The new value for the `colData` slot to set.}
}
\value{
Returns the `colData` slot when getting, and the modified `scmp` object when setting.
}
\description{
Set the Sparse Column Data of an scmp Object
}
