% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{cSparse<-,scmp-method}
\alias{cSparse<-,scmp-method}
\title{Set the Sparse Column Data of an scmp Object}
\usage{
\S4method{cSparse}{scmp}(object) <- value
}
\arguments{
\item{object}{An object of class `scmp`.}

\item{value}{The new value to set in the `colData` slot.}
}
\value{
Returns the modified `scmp` object.
}
\description{
This method allows setting the `colData` of the `sparse` slot of a `scmp` object.
}
