% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{cSparse}
\alias{cSparse}
\title{Get or set the sparse column data of a scmp object}
\usage{
cSparse(object, value = "missing")
}
\arguments{
\item{object}{An object of class `scmp`.}

\item{value}{The new value for the `colData` slot to set. 
This parameter is optional for getting the data.}
}
\value{
Returns the `colData` slot when getting. 
        Returns the modified `scmp` object when setting.
}
\description{
This function can be used to either retrieve or set the `colData` 
from the `sparse` slot of a `scmp` object.
}
