\name{data.abiotic}
\alias{data.abiotic}
\docType{data}
\title{ Gene expression data potato abiotic stress}
\description{

 \code{data.abiotic} contains gene expression of a time course microarray experiment where potato plants were submitted to 3 different abiotic stresses.

}
\usage{data(data.abiotic)}
\format{
  A data frame with 1000 observations on the following 36 variables.
  \describe{
    \item{\code{Control_3H_1}}{a numeric vector}
    \item{\code{Control_3H_2}}{a numeric vector}
    \item{\code{Control_3H_3}}{a numeric vector}
    \item{\code{Control_9H_1}}{a numeric vector}
    \item{\code{Control_9H_2}}{a numeric vector}
    \item{\code{Control_9H_3}}{a numeric vector}
    \item{\code{Control_27H_1}}{a numeric vector}
    \item{\code{Control_27H_2}}{a numeric vector}
    \item{\code{Control_27H_3}}{a numeric vector}
    \item{\code{Cold_3H_1}}{a numeric vector}
    \item{\code{Cold_3H_2}}{a numeric vector}
    \item{\code{Cold_3H_3}}{a numeric vector}
    \item{\code{Cold_9H_1}}{a numeric vector}
    \item{\code{Cold_9H_2}}{a numeric vector}
    \item{\code{Cold_9H_3}}{a numeric vector}
    \item{\code{Cold_27H_1}}{a numeric vector}
    \item{\code{Cold_27H_2}}{a numeric vector}
    \item{\code{Cold_27H_3}}{a numeric vector}
    \item{\code{Heat_3H_1}}{a numeric vector}
    \item{\code{Heat_3H_2}}{a numeric vector}
    \item{\code{Heat_3H_3}}{a numeric vector}
    \item{\code{Heat_9H_1}}{a numeric vector}
    \item{\code{Heat_9H_2}}{a numeric vector}
    \item{\code{Heat_9H_3}}{a numeric vector}
    \item{\code{Heat_27H_1}}{a numeric vector}
    \item{\code{Heat_27H_2}}{a numeric vector}
    \item{\code{Heat_27H_3}}{a numeric vector}
    \item{\code{Salt_3H_1}}{a numeric vector}
    \item{\code{Salt_3H_2}}{a numeric vector}
    \item{\code{Salt_3H_3}}{a numeric vector}
    \item{\code{Salt_9H_1}}{a numeric vector}
    \item{\code{Salt_9H_2}}{a numeric vector}
    \item{\code{Salt_9H_3}}{a numeric vector}
    \item{\code{Salt_27H_1}}{a numeric vector}
    \item{\code{Salt_27H_2}}{a numeric vector}
    \item{\code{Salt_27H_3}}{a numeric vector}
  }
}
\details{
This data set is part of a larger experiment in wich gene expression was monitored in 
both roots and leaves using a 11K cDNA potato chip. This example data set contains a ramdom 
subset of 1000 genes of the leave study.  
}

\references{
Rensink WA, Iobst S, Hart A, Stegalkina S, Liu J, Buell CR. Gene expression 
profiling of potato responses to cold, heat, and salt stress. Funct Integr 
Genomics. 2005 Apr 22.
}

\examples{
data(data.abiotic)

}
\keyword{datasets}
