% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpClass.R
\docType{class}
\name{designClass}
\alias{designClass}
\alias{designClass-class}
\title{Class "designClass"}
\description{
An S4 class to represent an design object with associated data.
This class contains three slots: predictor, groups.vector, and alloc.
}
\section{Slots}{

\describe{
\item{\code{predictor}}{A data frame containing the design matrix of dummies for fitting the GLM.}

\item{\code{groups.vector}}{A character vector specifying the experimental group to which
each variable belongs to.}

\item{\code{alloc}}{A data frame describing the experimental design. Rows must contain
cells and columns experiment descriptors. The matrix must be binarized.}
}}

\section{Validity}{

  Valid objects must have:
  \itemize{
    \item{predictor}{A valid data frame.}
    \item{groups.vector}{A valid character vector.}
    \item{alloc}{A valid data frame.}
  }
}

\keyword{classes}
