% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpShow.R
\name{.smsp_show}
\alias{.smsp_show}
\title{Show ScMaSigPro Object Information}
\usage{
.smsp_show(object)
}
\arguments{
\item{object}{An object of class \code{scMaSigProClass}.}
}
\description{
This method displays basic information about the ScMaSigPro object when the object
is printed in the console. The method is automatically called when the user writes
the name of the object in the console.
}
\keyword{internal}
