% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eDense,ScMaSigPro-method}
\alias{eDense,ScMaSigPro-method}
\title{Get eDense value for ScMaSigPro objects}
\usage{
\S4method{eDense}{ScMaSigPro}(object, value = "missing")
}
\arguments{
\item{object}{The ScMaSigPro object.}

\item{value}{Dummy parameter, not used.}
}
\value{
The eDense value from the ScMaSigPro object.
}
\description{
Method to get the eDense value from ScMaSigPro class objects.
}
