% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eDense,scmp-method}
\alias{eDense,scmp-method}
\title{Get eDense value for scmp objects}
\usage{
\S4method{eDense}{scmp}(object, value = "missing")
}
\arguments{
\item{object}{The scmp object.}

\item{value}{Dummy parameter, not used.}
}
\value{
The eDense value from the scmp object.
}
\description{
Method to get the eDense value from scmp class objects.
}
