% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eDense<-,ScMaSigPro,character,matrix-method}
\alias{eDense<-,ScMaSigPro,character,matrix-method}
\title{Set or Update Expression Data in Dense Slot}
\usage{
\S4method{eDense}{ScMaSigPro,character,matrix}(object, i = "bulk.counts") <- value
}
\arguments{
\item{object}{An \code{\link{ScMaSigPro}} object to be modified.}

\item{i}{The name of the dataset within the Dense slot to set or update.}

\item{value}{The expression data to set for the specified dataset.}
}
\value{
The modified \code{\link{ScMaSigPro}} object with the updated or new expression data
in the Dense slot.
}
\description{
`eDense<-` updates or sets new expression data within the Dense slot of an
\code{\link{ScMaSigPro}} object.
}
\seealso{
\code{\link{eDense}} for retrieving expression data from the Dense slot.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
