% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eDense<-,ScMaSigPro-method}
\alias{eDense<-,ScMaSigPro-method}
\title{Get Value for expression counts.}
\usage{
\S4method{eDense}{ScMaSigPro}(object, i) <- value
}
\arguments{
\item{object}{An object of class \code{\link{ScMaSigPro}}.}

\item{i}{Name of the Assay.}

\item{value}{The value to be set. (If Setting)}
}
\value{
Modified `ScMaSigPro` object.
}
\description{
`eDense<-` is a generic function for setting expression counts for
Sparse in an \code{\link{ScMaSigPro}}.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
