% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eDense<-,scmp-method}
\alias{eDense<-,scmp-method}
\title{Set eDense value for scmp objects (Not Implemented)}
\usage{
\S4method{eDense}{scmp}(object, i) <- value
}
\arguments{
\item{object}{The scmp object.}

\item{i}{object name}

\item{value}{The value to set.}
}
\value{
None
}
\description{
Method to set the eDense value for scmp class objects.
This method is currently not implemented.
}
