% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eDense<-}
\alias{eDense<-}
\title{Set eDense value for an object}
\usage{
eDense(object, i) <- value
}
\arguments{
\item{object}{The object to be modified.}

\item{i}{object name}

\item{value}{The value to be set for eSparse.}
}
\value{
None
}
\description{
`eDense<-` is a generic function for setting eDense value in an object.
Currently, this functionality is not implemented for all object types.
}
