% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eDense}
\alias{eDense}
\title{Get Value for expression counts.}
\usage{
eDense(object, value = "missing")
}
\arguments{
\item{object}{An object of class \code{\link{ScMaSigPro}}.}

\item{value}{The value to be set. (If Setting)}
}
\value{
The expression matrix for Dense slot.
}
\description{
`eDense<-` is a generic function for getting expression counts for Dense
in an \code{\link{ScMaSigPro}}.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
