% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edesignClass.R
\docType{class}
\name{edesignClass}
\alias{edesignClass}
\alias{edesignClass-class}
\title{Class "edesignClass"}
\description{
An S4 class to represent an edesign object with associated data.
This class contains three slots: dis, groups.vector, and edesign.
}
\section{Slots}{

\describe{
\item{\code{dis}}{A data frame containing the design matrix of dummies for fitting the GLM.}

\item{\code{groups.vector}}{A character vector specifying the experimental group to which
each variable belongs to.}

\item{\code{edesign}}{A data frame describing the experimental design. Rows must contain
cells and columns experiment descriptors. The matrix must be binarized.}

\item{\code{poly_degree}}{Integer with the polynomial degree to fit the regression. 1 
specifies a linear regression, 2 a quadratic regression, etc.}
}}

\section{Validity}{

  Valid objects must have:
  \itemize{
    \item{dis}{A valid data frame.}
    \item{groups.vector}{A valid character vector.}
    \item{edesign}{A valid data frame.}
  }
}

\keyword{classes}
