% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_discretize.R
\name{entropy_discretize}
\alias{entropy_discretize}
\title{Find optimal number of pseudotime bins}
\usage{
entropy_discretize(
  scmpObject,
  pseudotime_colname = scmpObject@addParams@pseudotime_colname,
  path_colname = scmpObject@addParams@path_colname,
  bin_method = "Sturges",
  drop.fac = 0.5,
  verbose = TRUE,
  binning = "universal",
  bin_pseudotime_colname = scmpObject@addParams@bin_pseudotime_colname
)
}
\arguments{
\item{scmpObject}{object of Class scMaSigPro. See \code{\link{scMaSigProClass}}
for more details.}

\item{pseudotime_colname}{Name of the column in `cell.metadata` storing
Pseudotime values. Generated using `colData` from the \pkg{SingleCellExperiment}
package. (Default is "Pseudotime").}

\item{path_colname}{Name of the column in `cell.metadata` storing information
for Path. Generated using `colData` from the \pkg{SingleCellExperiment}
package. (Default is `path_prefix`).}

\item{bin_method}{A character string specifying the method to use in order to
estimate the optimal number of bins. Available options: "Freedman.Diaconis",
"Sqrt", "Sturges", "Rice", "Doane", and "Scott.Normal". See \code{\link{estBinSize}}
for more details. (Default = "Sturges").}

\item{drop.fac}{A numeric value specifying the factor by which to decrease the
number of bins if the initial binning results in too many bins. (Default = 0.5).}

\item{verbose}{Print detailed output in the console. (Default is TRUE)}

\item{binning}{A character string. When set to "individual", the bins are calculated
per path iteratively. Options: "universal", "individual. (Default = "universal").}

\item{bin_pseudotime_colname}{Name of the column to store the computed Pseudotime
bins.}
}
\value{
A data.frame that contains the original data plus additional columns:
- 'bin' : Name of the bin
- 'bin_size' : Size of the bin
- 'binned_time' : Interval range of each bin
This function returns the merged data.frame with new discretized
pseudotime_colname, preserving the original rownames.
}
\description{
`entropy_discretize()` discretizes a continuous time series column into bins
of equal size using entropy-based binning method. It automatically calculates
the optimal number of bins using one of the supported methods. The bin sizes
are also calculated and merged with the input cell_metadata.
}
\details{
This function performs the following steps:
- Adds a new column 'cell' to the input data.frame which copies the row names.
- Extracts the time series data from the specified column of the input data.frame.
- Calculates the optimal number of bins using the specified method.
- Prints the estimated number of bins if verbose is set to TRUE.
- Discretizes the time series data into bins using the entropy-based binning method.
- Merges the original data.frame with the new binned time series data.
- Removes the 'cell' column and sets the row names back to the original row names of the input data.frame.
- Returns the merged data.frame.
}
\examples{
\dontrun{
entropy_discretize(
  cell_metadata = data.frame, pseudotime_colname = "time",
  bin_method = "Sturges", drop.fac = 0.5, verbose = TRUE
)
}

}
\seealso{
\code{\link{estBinSize}}, \code{\link{discretize}}, \code{\link{create_range}}
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
