% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eSparse,ScMaSigPro-method}
\alias{eSparse,ScMaSigPro-method}
\title{Get eSparse value for ScMaSigPro objects}
\usage{
\S4method{eSparse}{ScMaSigPro}(object, value = "missing")
}
\arguments{
\item{object}{The ScMaSigPro object.}

\item{value}{Dummy parameter, not used.}
}
\value{
The eSparse value from the ScMaSigPro object.
}
\description{
Method to get the eSparse value from ScMaSigPro class objects.
}
