% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eSparse,scmp-method}
\alias{eSparse,scmp-method}
\title{Get eSparse value for scmp objects}
\usage{
\S4method{eSparse}{scmp}(object, value = "missing")
}
\arguments{
\item{object}{The scmp object.}

\item{value}{Dummy parameter, not used.}
}
\value{
The eSparse value from the scmp object.
}
\description{
Method to get the eSparse value from scmp class objects.
}
