% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eSparse<-,ScMaSigPro,character,matrix-method}
\alias{eSparse<-,ScMaSigPro,character,matrix-method}
\title{Set Expression Counts in Sparse Slot}
\usage{
\S4method{eSparse}{ScMaSigPro,character,matrix}(object, i = "counts") <- value
}
\arguments{
\item{object}{An \code{\link{ScMaSigPro}} object to be modified.}

\item{i}{The name of the assay for which to set or update expression counts.
Must be provided as a character string.}

\item{value}{The new expression data to set for the specified assay.}
}
\value{
The modified \code{\link{ScMaSigPro}} object with the updated or new expression data
in the Sparse slot.
}
\description{
Sets or updates expression counts in the Sparse slot of an 
\code{\link{ScMaSigPro}} object.
}
\seealso{
\code{\link{eSparse}} for retrieving expression counts from the Sparse slot.
}
