% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eSparse<-}
\alias{eSparse<-}
\title{Set value for expression counts.}
\usage{
eSparse(object, i) <- value
}
\arguments{
\item{object}{An object of class \code{\link{ScMaSigPro}}.}

\item{i}{Name of the assay.}

\item{value}{The value to be set.}
}
\value{
Modified `ScMaSigPro` object.
}
\description{
`eSparse<-` is a generic function for setting expression counts for Sparse
in an \code{\link{ScMaSigPro}}.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
