% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScMaSigProClass.R
\docType{class}
\name{Estimates}
\alias{Estimates}
\alias{Estimates-class}
\title{Estimates}
\description{
S4 class to store results of the polynomial term selection from full model.
}
\section{Slots}{

\describe{
\item{\code{sol}}{A data frame for summary results of the stepwise regression.
For each selected gene the following values are given:
  \itemize{
    \item{p-value}{of the regression ANOVA.}
    \item{R-squared}{of the model.}
    \item{p-value}{of the regression coefficients of the selected variables.}
  }}

\item{\code{coefficients}}{A data frame containing regression coefficients for the adjusted models.}

\item{\code{group.coeffs}}{A matrix with the coefficients of the implicit models of each experimental group.}

\item{\code{t.score}}{A data frame containing tscores for each covariate in polynomial glm.}

\item{\code{path}}{A character vector containing the branching path.}

\item{\code{influ.info}}{A matrix with genes containing influencial data.}
}}

\keyword{classes}
