% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.intervals.R
\name{extract.intervals}
\alias{extract.intervals}
\title{Extract Intervals}
\usage{
extract.intervals(time.vector, nBins = 1, bin, bin.size, lbound, ubound)
}
\arguments{
\item{time.vector}{description}

\item{nBins}{description}

\item{bin}{description}

\item{bin.size}{description}

\item{lbond}{description}

\item{ubond}{description}
}
\description{
Extract Intervals
}
\keyword{internal}
