% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fitting.R
\name{extract_fitting}
\alias{extract_fitting}
\title{Extract Fitting}
\usage{
extract_fitting(
  reg,
  lmf,
  model.glm.0,
  dis,
  family,
  name,
  vars.in,
  alfa,
  influ.info
)
}
\arguments{
\item{reg}{description}

\item{lmf}{description}

\item{model.glm.0}{description}

\item{dis}{description}

\item{family}{description}

\item{name}{description}

\item{vars.in}{description}

\item{alfa}{description}

\item{influ.info}{description}
}
\description{
Extract Fitting
}
\keyword{internal}
