% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_functions.R
\name{extract_info}
\alias{extract_info}
\title{Extract Information `.ScMaSigPro_show`}
\usage{
extract_info(data, return_type = "avg_bin_size", bin_size_col, path_col)
}
\arguments{
\item{data}{A dataframe.}

\item{return_type}{Average Bin Size or Number of Bins.}

\item{bin_size_col}{Column name with the bins.}

\item{path_col}{Column name with the path information.}
}
\description{
Extract details for the console cat of `.ScMaSigPro_show`.
}
\keyword{internal}
