% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_monocle3.R
\name{extract_monocle3_components}
\alias{extract_monocle3_components}
\title{Extract Monocle3 Components}
\usage{
extract_monocle3_components(cds, reduction_method = "umap", verbose = TRUE)
}
\arguments{
\item{cds}{A Monocle3 CellDataSet (CDS).}

\item{reduction_method}{The dimensionality reduction method used. Default is "umap".}

\item{verbose}{Logical. If TRUE, print the identified root points and information about the longest and shortest paths. Default is TRUE.}
}
\value{
A list containing:
\itemize{
 \item{cds}{The updated Monocle3 CellDataSet.}
 \item{y_to_cells}{A dataframe containing the Y coordinates of the cells.}
 \item{endpoints}{The endpoint nodes of the minimum spanning tree.}
 \item{root}{The root nodes of the minimum spanning tree.}
 \item{num_cells_per_path}{A dataframe containing the number of cells for each path.}
}
}
\description{
This function extracts the relevant components from a Monocle3 CellDataSet (CDS). These components include the
principal points, root cells, minimum spanning tree, network matrix, pseudotime, number of cells per path, and cell metadata.
It then updates the CDS with the endpoint data and the updated cell metadata.
}
\details{
The function first extracts the principal points and root cells from the CDS. It then calculates the minimum spanning tree
and the network matrix, which contains the cells along each path from root to endpoint.
The function also calculates the pseudotime for each path and the number of cells per path.
Finally, it updates the CDS with the endpoint data and the updated cell metadata.
}
\examples{
\dontrun{
# Assuming 'cds' is a Monocle3 CellDataSet (CDS)
result <- extract_monocle3_components(cds, reduction_method = "umap", verbose = TRUE)
}

}
\seealso{
\code{\link[monocle3]{principal_graph_aux}}, \code{\link[monocle3]{principal_graph}}, \code{\link[igraph]{shortest_paths}}, \code{\link[monocle3]{pseudotime}}
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
