% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_slingshot_components.R
\name{extract_slingshot_components}
\alias{extract_slingshot_components}
\title{Extract Slingshot components}
\usage{
extract_slingshot_components(
  sling.sce,
  reduction_method = "umap",
  verbose = TRUE
)
}
\arguments{
\item{sling.sce}{A SingleCellExperiment object processed by Slingshot.}

\item{reduction_method}{A string indicating the dimensionality reduction method to use. Default is "umap".}

\item{verbose}{A logical value indicating whether to print detailed messages. Default is TRUE.}
}
\value{
A list containing the extracted sling curves, sling time, and sling dimensions, each as a dataframe.
}
\description{
This function extracts the sling curves, sling time, and sling dimensions from a given SingleCellExperiment object
processed by the Slingshot package. The function also allows choosing the dimensionality reduction method to use.
}
\examples{
\dontrun{
result <- extract_slingshot_components(sce)
}
}
