% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableDS.R
\name{f3}
\alias{f3}
\title{Merge multiple cluster numbers into a single string with '&'}
\usage{
f3(x)
}
\arguments{
\item{x}{A numeric vector of cluster numbers}
}
\value{
A string with cluster numbers separated by '&'
}
\description{
This function takes a vector of cluster numbers and merges them into a single string
by separating the numbers with '&'.
}
\examples{
x <- c(1, 2, 4, NA, NA)
f3(x)
}
