% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fromListWithNames}
\alias{fromListWithNames}
\title{Convert List of Named Vectors to Data Frame with Row Names for UpSetR}
\usage{
fromListWithNames(input)
}
\arguments{
\item{input}{A list of named vectors to be converted to a data frame compatible with UpSetR.}
}
\value{
A data frame with binary values indicating the presence (1) or absence (0) of each element
in the sets, and row names corresponding to the unique elements.
}
\description{
This function converts a list of named vectors to a data frame compatible with UpSetR,
retaining the row names corresponding to the unique elements (genes) present in the input list.
}
\keyword{internal}
