% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.features.R
\name{get.features}
\alias{get.features}
\title{Get Features from scMaSigProClass Object}
\usage{
get.features(
  scmpObj,
  query = "unique",
  unique.group = "Path1",
  union.ref = "Path1",
  union.target = "Path2vsPath1",
  rsq = 0.7,
  Q = scmpObj@addParams@Q,
  vars = "groups",
  significant.intercept = "dummy",
  term.Q = 0.05,
  includeInflu = TRUE,
  unique.trend = "any",
  union.ref.trend = "any",
  union.target.trend = "any"
)
}
\arguments{
\item{scmpObj}{An object of class 'scMaSigProClass'.}

\item{query}{A string specifying the type of query.
Valid options are 'unique' or 'union'.
Default is 'unique'.}

\item{unique.group}{A string specifying the group to consider.
Default is 'Path1'.}
}
\value{
Depending on the query, the function returns either a vector of
        unique features (when query is 'unique') or a vector of features
        common to all groups (when query is 'union').
}
\description{
This function extracts features from an object of class 'scMaSigProClass'
based on a specified query and group. The function supports two types of
queries: 'unique' or 'union', and allows for the selection of a specific
group from the available groups in the object.
}
\examples{
# Assuming 'scmp' is an object of class 'scMaSigProClass':
features_unique <- get.features(scmp, query = "unique", group = "Group1")
features_union <- get.features(scmp, query = "union")

}
