% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_os}
\alias{get_os}
\title{Get Identifying the OS from R.}
\usage{
get_os()
}
\value{
A string indicating the operating system: either 'osx', 'linux', or a
lower-case version of `Sys.info()['sysname']` if the system is neither OS X
nor Linux.
}
\description{
This function returns the name of the operating system the R session is
running on. It differentiates between 'osx', 'linux', and other systems.
Visit https://www.r-bloggers.com/
}
\author{
Will
}
\keyword{internal}
