% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLowestAICModel.R
\name{getLowestAICModel}
\alias{getLowestAICModel}
\title{Get Model with Lowest AIC Function}
\usage{
getLowestAICModel(model_list)
}
\description{
Description:
This function takes a list of fitted models and returns the name of the model
with the lowest AIC (Akaike Information Criterion) value.
}
\details{
Usage:
getLowestAICModel(model_list)

Arguments:
  \describe{
    \item{model_list}{A list of fitted models for which you want to find the 
                     one with the lowest AIC. Each element of the list should be 
                     a fitted model object that supports the AIC function.}
  }

Details:
The function computes the AIC values for all the models in the input list 
using the `AIC` function. It then finds the index of the model with the 
lowest AIC value and retrieves its name.

Value:
The function returns the name of the model with the lowest AIC value as a character.
}
\examples{
# Fit models and create a list
model1 <- lm(mpg ~ wt, data = mtcars)
model2 <- lm(mpg ~ wt + hp, data = mtcars)
model3 <- lm(mpg ~ wt + hp + qsec, data = mtcars)
model_list <- list(model1, model2, model3)

# Get the model with the lowest AIC
lowest_aic_model <- getLowestAICModel(model_list)
print(lowest_aic_model)

}
\keyword{internal}
