% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_inspection.R
\name{glm_inspection}
\alias{glm_inspection}
\title{Visually Inspect Model Fit}
\usage{
glm_inspection(
  glm.model,
  plot.type = "fitted_vs_pearson",
  point.color = "blue",
  point.shape = 3,
  bar_width = 0.05,
  point.alpha = 0.8,
  ref.line.y.intercept = 0,
  func = "glm",
  input.data = NULL
)
}
\arguments{
\item{glm.model}{model from the output of \link[stats]{glm}.}

\item{plot.type}{Plot-type. Currentry supported options are
_fitted_vs_pearson_, _pearson_vs_obs_, _fitted_vs_obs_}

\item{point.color}{Point colour.}

\item{point.shape}{Point shape.}

\item{bar_width}{bar width of the cook's distance plot}

\item{point.alpha}{Point alpha.}

\item{ref.line.y.intercept}{Y intercept of the reference line.}

\item{func}{Which function is used}

\item{input.data}{No information}
}
\description{
Visually Inspect Model Fit
}
