% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenize_bins.R
\name{homogenize_bins}
\alias{homogenize_bins}
\title{Homogenize Bins of a Dataframe}
\usage{
homogenize_bins(
  df,
  l_bound_col,
  u_bound_col,
  bin_size_col,
  binPTime_col,
  verbose = TRUE
)
}
\arguments{
\item{df}{A dataframe containing the bin data.}

\item{l_bound_col}{The name of the lower bound column as a string.}

\item{u_bound_col}{The name of the upper bound column as a string.}

\item{bin_size_col}{The name of the bin size column as a string.}

\item{binPTime_col}{The name of the column with Pseudotime}
}
\value{
A dataframe with adjusted bins.
}
\description{
Adjust the bins of a dataframe based on a strategy that checks if the bin size is too small or too large
relative to the mean and standard deviation of the bin sizes. If a bin is too small, it will be merged 
with the previous bin. If it's too large, it will be split into two equal parts.
}
\keyword{internal}
