% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.rank.R
\name{i.rank}
\alias{i.rank}
\title{Ranks a vector to index}
\usage{
i.rank(x)
}
\arguments{
\item{x}{The input vector that will be ranked.}
}
\value{
Vector of ranked values.
}
\description{
\code{i.rank} ranks the values in a vector to successive values. Ties are given the same value.
}
\details{
Ranks a vector to index

\code{i.rank} ranks the values in a vector to successive values. Ties are given the same value.
}
\examples{
i.rank(c(1, 1, 1, 3, 3, 5, 7, 7, 7))

}
\seealso{
\code{\link{rank}},\code{\link{order}} Functions related to ranking and ordering vectors.
}
\author{
Ana Conesa and Maria Jose Nueda (mj.nueda@ua.es)
}
\keyword{arith}
