% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsoModel.R
\name{IsoModel}
\alias{IsoModel}
\title{IsoModel}
\usage{
IsoModel(
  data,
  gen,
  design = NULL,
  Q = 0.05,
  min.obs = 6,
  minorFoldfilter = NULL,
  counts = FALSE,
  family = NULL,
  theta = 10,
  epsilon = 1e-05
)
}
\arguments{
\item{data}{A matrix holding isoform expression data. Isoforms should be in rows and experimental conditions in columns.}

\item{gen}{A vector with the name of the gene each isoform is tied to.}

\item{design}{The design matrix for the regression fit, like that generated by the `make.design.matrix` function. Default is NULL.}

\item{Q}{The significance level. Default is 0.05.}

\item{min.obs}{Cases having less than this count of true numerical values will be removed from the analysis. The minimum value to estimate the model is (degree+1)xGroups+1. Default is 6.}

\item{minorFoldfilter}{Fold expression difference between minor isoforms and the most expressed isoform to remove minor isoforms from the analysis. Default is NULL.}

\item{counts}{A logical indicating whether your data are counts. Default is FALSE.}

\item{family}{The distribution function to be utilized in the glm model. Must be specified as a function: gaussian(), poisson(), negative.binomial(theta) etc. If NULL, family will be negative.binomial(theta) when counts=TRUE or gaussian() when counts=FALSE.}

\item{theta}{Theta parameter for negative.binomial family. Default is 10.}

\item{epsilon}{Argument to pass to `glm.control`, convergence tolerance in the iterative process to estimate de glm model. Default is 1e-05.}
}
\value{
A list containing:
* `data`: The input data matrix used in the following steps.
* `gen`: The input gen vector used in the following steps.
* `design`: The input design matrix used in the following steps.
* `DSG`: Names of the selected genes: Differentially Spliced Genes.
* `pvector`: p.vector output of isoforms that belong to selected.genes.
* `Tfit`: Tfit output of isoforms that belong to selected.genes.
}
\description{
Detect Genes with Isoforms Expressing Differently in Time Course Experiments
}
\details{
The function `IsoModel` carries out a model comparison for each gene to pinpoint genes with varying trends in time course experiments, and implements maSigPro to the Isoforms associated with selected genes.


The vectors `rownames(design)` and `colnames(data)` must be identical and represent experimental condition names.
The `rownames(data)` should contain unique isoform IDs.
The `colnames(design)` are the given names for the variables in the regression model.
}
\examples{
data(ISOdata)
data(ISOdesign)
mdis <- make.design.matrix(ISOdesign)
MyIso <- IsoModel(data = ISOdata[, -1], gen = ISOdata[, 1], design = mdis, counts = TRUE)
Myget <- getDS(MyIso)
see <- seeDS(Myget, cluster.all = FALSE, k = 6)
table <- tableDS(see)
table$IsoTable
}
\references{
Nueda, M.J., Martorell, J., Marti, C., Tarazona, S., Conesa, A. 2018. Identification and visualization of differential isoform expression in RNA-seq time series. Bioinformatics. 34, 3, 524-526.
Nueda, M.J., Tarazona, S., Conesa, A. 2014. Next maSigPro: updating maSigPro bioconductor package for RNA-seq time series. Bioinformatics, 30, 2598-602.
Conesa, A., Nueda M.J., Ferrer, A., Talon, T. 2006. maSigPro: a Method to Identify Significant Differential Expression Profiles in Time-Course Microarray Experiments. Bioinformatics 22, 1096-1102.
}
\seealso{
`p.vector`, `T.fit`
}
\author{
Maria Jose Nueda, \email{mj.nueda@ua.es}
}
