% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.bulk.count.R
\name{make.pseudobulk.counts}
\alias{make.pseudobulk.counts}
\title{make.pseudobulk.counts}
\usage{
make.pseudobulk.counts(
  counts,
  cluster_member_col = "cluster.members",
  bin_col = "bin",
  pseudo_bulk_profile,
  cluster.count.by = "sum"
)
}
\arguments{
\item{counts}{A data.frame. The count data from single cell RNA-seq experiment.}

\item{cluster_member_col}{A character string (default = "cluster.members"). The name of the column in the pseudo_bulk_profile that contains the cluster members.}

\item{bin_col}{A character string (default = "bin"). The name of the column in the pseudo_bulk_profile that contains the bin identifiers.}

\item{pseudo_bulk_profile}{A data.frame. The pseudo bulk profile created by make_pseudobulk_design function.}

\item{cluster.count.by}{A character string (default = "sum"). The method to use to aggregate counts within each cluster. Possible values are "mean" or "sum".}
}
\value{
A data.frame. The data frame includes pseudo bulk counts with each row being a gene and each column being a bin.
}
\description{
This function creates a dataframe of pseudo bulk counts from single cell counts. It does this by either taking the mean or sum of counts across clusters in each bin, depending on the specified method.
}
\details{
The function operates by iterating over each row of the pseudo_bulk_profile. For each bin, it identifies the cells that belong to the bin and selects their counts from the counts data frame. It then calculates the mean or sum of these counts (depending on the specified method), and adds these to a new data frame of pseudo bulk counts. The result is a pseudo bulk counts data frame where each row is a gene and each column is a bin.
}
\examples{
\dontrun{
make.pseudobulk.counts(counts = sc_counts, 
                 pseudo_bulk_profile = pb_profile,
                 cluster.count.by = "mean")
}

}
