% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.bulk.design.R
\name{make.pseudobulk.design}
\alias{make.pseudobulk.design}
\title{make.pseudobulk.design}
\usage{
make.pseudobulk.design(design.file, pathCol, binnedCol = "binnedTime")
}
\arguments{
\item{design.file}{A data.frame. The design file containing the binned data.}

\item{pathCol}{A character string. The name of the column in the design file that contains the path data.}

\item{binnedCol}{A character string (default = "binnedTime"). The name of the column in the design file that contains the binned data.}
}
\value{
A data.frame containing the pseudobulk profile. The data frame includes the following columns:
- 'binnedTime': The time bin
- 'cluster.members': The cells that fall into the bin
- 'bin': The bin identifier
- 'path': The path identifier
- 'bin.size': The size of the bin (number of cells in the bin)
- Columns for each path in paths.vector, with binary values indicating whether the row belongs to the path
}
\description{
This function creates a pseudobulk profile data frame from the design file. It groups cells into bins for each path from a given set of paths and calculates the bin size.
}
\details{
This function operates by iterating over the specified paths. For each path, it:
- Filters the design file to only include cells that belong to the path
- Groups the cells into bins based on the binned time column
- Calculates the size of each bin
- Appends the resulting data frame to the pseudobulk profile

After processing all paths, the function adds binary columns for each path to the pseudobulk profile.
}
\examples{
\dontrun{
make.pseudobulk.design(design.file = df, pathCol = "path1", binnedCol = "binnedTime")
}

}
\seealso{
\code{\link{calc_bin_size}}
}
