% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScMaSigProClass.R
\docType{class}
\name{MatrixDesign}
\alias{MatrixDesign}
\alias{MatrixDesign-class}
\title{MatrixDesign}
\description{
S4 class to store Branching Path Assignments and Prediction Matrix
}
\section{Slots}{

\describe{
\item{\code{predictor_matrix}}{A matrix containing independent variables for
model fitting}

\item{\code{groups.vector}}{A character vector specifying the branching path for each
term of the polynomial GLM.}

\item{\code{assignment_matrix}}{A matrix containing binary assignment to branching paths
Additionally has two columns for assignment of binned Pseudotime and
replicate.}
}}

\keyword{classes}
