% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableDS.R
\name{MayorIso}
\alias{MayorIso}
\title{Identify the major isoform for a gene}
\usage{
MayorIso(zz)
}
\arguments{
\item{zz}{A data frame containing gene expression values across clusters}
}
\value{
A numeric vector indicating the major isoform (1) and minor isoforms (0)
}
\description{
This function takes a gene profile and identifies the major isoform based on the highest sum
of expression values across all clusters.
}
\examples{
data.clust <- data.frame(Cluster1 = c(1, 0, 3), Cluster2 = c(2, 5, 1))
MayorIso(data.clust)
}
