% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScMaSigProClass.R
\docType{class}
\name{Misc}
\alias{Misc}
\alias{Misc-class}
\title{Misc}
\description{
S4 class to store miscellaneous data.
}
\section{Slots}{

\describe{
\item{\code{r_version}}{R version used to instanciate the object.}

\item{\code{os}}{Name of the OS in which object was instanciated.}

\item{\code{package_version}}{Version of the package used to instanciate the object.}

\item{\code{last_modified}}{Last modified date of the object.}

\item{\code{last_operation}}{last operation performed on the object.}

\item{\code{date_created}}{Date of object creation.}
}}

\keyword{classes}
