\name{NBdata}
\alias{NBdata}
\docType{data}
\title{ RNA-Seq dataset example}
\description{

 \code{NBdata} contains a subset of a bigger normalized negative binomial simulated dataset.

}
\usage{data(NBdata)}
\format{
  A data frame with 100 observations on 36 numeric variables.
}
\details{
This dataset is part of a larger simulated and normalized dataset with 2 experimental groups, 6 time-points and 3 replicates.
Simulation has been done by using a negative binomial distribution.
The first 20 genes are simulated with changes among time.
}


\examples{
data(NBdata)

}
\keyword{datasets}
