% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpClass.R
\name{newScMaSigPro}
\alias{newScMaSigPro}
\title{Create scMaSigProClass Object}
\usage{
newScMaSigPro(
  SingleCellExperiment,
  covariate,
  parameters,
  pVector,
  tFit,
  sigGenes
)
}
\arguments{
\item{SingleCellExperiment}{An instance of \code{SingleCellExperiment}.}

\item{covariate}{An instance of \code{covariateClass}.}

\item{parameters}{An instance of \code{parameterClass}.}

\item{pVector}{An instance of \code{pVectorClass}.}

\item{tFit}{An instance of \code{tFitClass}.}

\item{sigGenes}{An instance of \code{sigGeneClass}.}
}
\value{
An instance of \code{scMaSigProClass}.
}
\description{
A constructor function to create an instance of \code{scMaSigProClass}.
}
