% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_helpers.R
\name{optimize_bin_max}
\alias{optimize_bin_max}
\title{Homogenize Bins of a Dataframe}
\usage{
optimize_bin_max(
  bin_table,
  max_allowed,
  verbose = TRUE,
  time_vector,
  lbound,
  ubound,
  bin,
  bin.size,
  method,
  drop
)
}
\arguments{
\item{bin_table}{A dataframe containing the bin data.}

\item{max_allowed}{The maximum allowed size for a bin.}

\item{verbose}{Logical; if TRUE, prints detailed output.}

\item{time_vector}{A vector representing time or a sequence that bins refer to.}

\item{lbound}{The name of the lower bound column in `bin_table`.}

\item{ubound}{The name of the upper bound column in `bin_table`.}

\item{bin}{The name of the bin identifier column in `bin_table`.}

\item{bin.size}{The name of the bin size column in `bin_table`.}

\item{method}{The method for handling small bins: 'merge' to merge with previous or next bin,
'drop' to remove small bins, or 'ignore' to leave small bins as they are.}

\item{drop}{The threshold below which a bin is considered too small and subject to the method.}
}
\value{
A dataframe with adjusted bins. The structure of the dataframe will be the same as `bin_table`
with updated values for the bin size and bounds.
}
\description{
Adjust the bins of a dataframe based on a specified maximum bin size threshold.
It checks if the bin size is too large relative to this threshold.
If a bin is too large, it will be split into two equal parts until all bins are
below the maximum allowed size. Optionally, bins too small can be merged with
adjacent bins based on the provided method.
}
\keyword{internal}
