% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpClass.R
\docType{class}
\name{parameterClass}
\alias{parameterClass}
\title{parameterClass}
\description{
A class to represent the analysis parameters for ScMaSigPro.
}
\section{Slots}{

\describe{
\item{\code{path.col}}{A character vector specifying the column name for the path data.}

\item{\code{time.col}}{A character vector specifying the column name for the time data.}

\item{\code{poly.order}}{An integer specifying the polynomial order for the time covariate.}

\item{\code{assay.name}}{A character vector specifying the assay name.}

\item{\code{p.vector.sig}}{A numeric vector specifying the significance level for p-vector.}

\item{\code{m.t.c.method}}{A character vector specifying the method for multiple testing correction.}

\item{\code{min.counts}}{A numeric vector specifying the minimum counts for genes.}

\item{\code{model.type}}{A character vector specifying the model type.}

\item{\code{model.control}}{A list specifying the model control parameters.}

\item{\code{covar.sig}}{A numeric vector specifying the significance level for covariate modeling.}
}}

