% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScMaSigProClass.R
\docType{class}
\name{ParameterConfig}
\alias{ParameterConfig}
\alias{ParameterConfig-class}
\title{ParameterConfig}
\description{
S4 Class to hold parameters used in the analysis. It is constantly updated
asthe analysis progresses.
}
\section{Slots}{

\describe{
\item{\code{bin_ptime_col}}{A character string representing the column name
for binned Pseudotime values in 'Dense' data. See `colData` from the
\pkg{SingleCellExperiment} package.}

\item{\code{path_prefix}}{Update Description..}

\item{\code{root_label}}{Update Description..}

\item{\code{ptime_col}}{A character string representing the column name
for inferred Pseudotime values in 'Sparse' data. See `colData` from the
\pkg{SingleCellExperiment} package.}

\item{\code{bin_method}}{A character string representing the algorithm used for
binning.}

\item{\code{path_col}}{A character string representing the column name for branching
path assignment in 'Sparse' or 'Dense'data. See `colData` from the
\pkg{SingleCellExperiment} package.}

\item{\code{bin_col}}{A character string representing the name of the column in which
bin labels are stored.}

\item{\code{bin_size_col}}{A character string representing the name of the column in
which bin sizes per bin are stored.}

\item{\code{bin_mem_col}}{A character string representing the name of the column in
which cells per bin are stored.}

\item{\code{anno_col}}{A character string representing the column name for cell level
metadata containing cell level annotations. (Default is "cell_type").}

\item{\code{g}}{Update Description..}

\item{\code{p_value}}{Significance Level.}

\item{\code{min_na}}{Minimum values needed to estimate the model.}

\item{\code{mt_correction}}{A character string specifying the p-value correction
method.}

\item{\code{epsilon}}{Model convergence tolerance.}

\item{\code{selection_method}}{A character string specifying the method for stepwise
regression.}

\item{\code{offset}}{A logical value specifying whether to use offset during fitting.}

\item{\code{log_offset}}{A logical value specifying whether to take the logarithm of
the offsets.}

\item{\code{max_it}}{Integer. Maximum number of iterations to fit the model.}

\item{\code{poly_degree}}{Integer with the polynomial degree to fit the regression. 1}

\item{\code{distribution}}{Distribution used}

\item{\code{cluster_method}}{Description}

\item{\code{use_dim}}{description}

\item{\code{fill_na}}{description}
}}

\keyword{classes}
