% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{pathAssign<-}
\alias{pathAssign<-}
\title{Replacement method for pathAssign.}
\usage{
pathAssign(object) <- value
}
\arguments{
\item{object}{An object of class \code{\link{ScMaSigPro}}.}

\item{value}{The value of the branch Assignment Matrix.}
}
\value{
Modified `ScMaSigPro` object with new Assignment Matrix.
}
\description{
Replacement method for pathAssign.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
