% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_counts.R
\name{pb_counts}
\alias{pb_counts}
\title{Create Pseduo-bulk Counts}
\usage{
pb_counts(
  scmpObject,
  bin_members_colname = scmpObject@param@bin_members_colname,
  bin_colname = scmpObject@param@bin_colname,
  assay_name = "counts",
  cluster_count_by = "sum"
)
}
\arguments{
\item{scmpObject}{object of Class scMaSigPro. See \code{\link{scmp}}
for more details.}

\item{bin_members_colname}{Column name in the 'compressed_cell_metadata'
storing information about the members of the bins. (Default is 'scmp_bin_members').}

\item{bin_colname}{Column name in the 'compressed_cell_metadata'
storing information about the bin labels. (Default is 'scmp_bin').}

\item{assay_name}{Name of the Assay in the assay_name object from which retrieve the counts.
(Default = "counts").}

\item{cluster_count_by}{A character string specifying the method to use to
aggregate counts within each cluster. Available options are 'mean' or 'sum'. (Default = "sum").}
}
\value{
A data.frame. The data frame includes pseudo bulk counts with each row being
a gene and each column being a bin.
}
\description{
`pb_counts()` creates a dataframe of pseudo bulk counts from single
cell counts. It does this by either taking the mean or sum of counts across clusters
in each bin, depending on the specified method.
}
\details{
The function operates by iterating over each row of the pseudo_bulk_profile.
For each bin, it identifies the cells that belong to the bin and selects their
counts from the counts data frame. It then calculates the mean or sum of these
counts (depending on the specified method), and adds these to a new data frame of pseudo bulk counts.
The result is a pseudo bulk counts data frame where each row is a gene and each column is a bin.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
\keyword{internal}
