% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loess_fit.R
\name{plot_loess_fit}
\alias{plot_loess_fit}
\title{Plot Spline Smoothers}
\usage{
plot_loess_fit(
  sce_obj,
  gene_name,
  log = F,
  assay_name = "TrueCounts",
  time_col = "Step",
  min_exp = 0,
  path_col = "Group",
  plt_subtitle = NULL,
  dfreedom = 3,
  span = 0.75
)
}
\arguments{
\item{sce_obj}{Object of the \link[SingleCellExperiment]{SingleCellExperiment} class.}

\item{gene_name}{Name of the gene to be plotted.}

\item{log}{Whether to plot log of the counts.}

\item{assay_name}{Name of the assay in the \link[SingleCellExperiment]{SingleCellExperiment} object.}

\item{time_col}{Column name having pseudotime information.}

\item{min_exp}{Minimum expression to plot.}

\item{path_col}{Column name having path information.}

\item{plt_subtitle}{Sub-title to add to the plot.}

\item{dfreedom}{Degrees of Freedom for loess fit.}

\item{span}{Smoothing parameter for the loess fit.}
}
\description{
Plot Spline Smoothers
}
