% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBinTile.R
\name{plotBinTile}
\alias{plotBinTile}
\title{Plot Bin Sizes Across Binned Time and Paths}
\usage{
plotBinTile(
  scmpObj,
  path_col = scmpObj@Parameters@path_col,
  bin_size_col = scmpObj@Parameters@bin_size_col,
  bin_ptime_col = scmpObj@Parameters@bin_ptime_col
)
}
\arguments{
\item{scmpObj}{A ScMaSigPro class object with an additional slot 'Dense' that
contains compression information.}

\item{path_col}{Name of the column in `cell.metadata` storing information
for Path.}

\item{bin_size_col}{A title of the barplot}

\item{bin_ptime_col}{description}
}
\value{
A tile plot made with `geom_tile()`, visualizing the bin sizes across
different binned time and paths.
}
\description{
This function generates plots to visualize the dense cell metadata from a ScMaSigPro
object. It produces tile plot to display the bin sizes across different
binned time intervals and paths.
}
