% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiagnostics.R
\name{plotDiagnostics}
\alias{plotDiagnostics}
\title{Plot Model Diagnostics}
\usage{
plotDiagnostics(scmpObj, feature_id, model = "optimized")
}
\arguments{
\item{scmpObj}{An object of class \code{\link{ScMaSigPro}}.}

\item{feature_id}{Name of the gene to be plotted.}

\item{model}{Type of model to be used for plotting. Can be either 'optimized',
'intercept'and 'full'. Default is 'optimized'.}
}
\value{
ggplot2 plot object.
}
\description{
Plots model residuals for the requested gene. Can be used for model
diagnostics of optimized model or full model.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
