% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIntersect.R
\name{plotIntersect}
\alias{plotIntersect}
\title{Generate UpSet Plot on Intersection of Significant Genes from scmpObject}
\usage{
plotIntersect(
  scmpObj,
  min_intersection_size = 2,
  keep_empty_groups = TRUE,
  width_ratio = 0.1,
  show_sets_size = FALSE
)
}
\arguments{
\item{scmpObj}{An object of class scmp}

\item{min_intersection_size}{minimal number of observations in an intersection
for it to be included.}

\item{keep_empty_groups}{whether empty sets should be kept (including sets
which are only empty after filtering by size)}

\item{width_ratio}{ratio of the overall set size width to intersection matrix
width.}

\item{show_sets_size}{the overall set sizes plot, e.g. from upset_set_size()
(FALSE to hide)}
}
\value{
An UpSet plot visualizing the intersections of significant genes across pathways.
}
\description{
Generate UpSet Plot on Intersection of Significant Genes from scmpObject
}
