% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrend.R
\name{plotTrend}
\alias{plotTrend}
\title{Plot Groups Function}
\usage{
plotTrend(
  scmpObj,
  feature_id,
  xlab = "Pooled Pseudotime",
  ylab = "Pseudobulk Expression",
  smoothness = 0.01,
  logs = TRUE,
  logType = "log",
  pseudoCount = 1,
  significant = TRUE
)
}
\arguments{
\item{scmpObj}{object of class scmpObj}

\item{feature_id}{A vector of maximum length 6 or a single feature of ID}

\item{xlab}{X-axis label. Default is "Pooled Pseudotime".}

\item{ylab}{Y-axis label. Default is "Pseudobulk Expression".}

\item{smoothness}{abc}

\item{logs}{Whether to plot log of counts}

\item{logType}{Log type required}

\item{pseudoCount}{add pseucount for log}

\item{significant}{description}
}
\value{
Generates a plot.
}
\description{
This function generates plots based on various parameters. It calculates the summary mode, colors, and other visual attributes to create a plot.
}
