% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{predictors,ScMaSigPro-method}
\alias{predictors,ScMaSigPro-method}
\title{Set or get the Predictor Matrix.}
\usage{
\S4method{predictors}{ScMaSigPro}(object, value = "missing")
}
\arguments{
\item{object}{An object of class \code{\link{ScMaSigPro}}.}

\item{value}{The value of the Predictors Matrix.}
}
\value{
Modified `ScMaSigPro` object with new Predictor Matrix.
}
\description{
Set or get the Predictor Matrix.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
