% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{predictors<-}
\alias{predictors<-}
\title{Replacement method for predictors.}
\usage{
predictors(object) <- value
}
\arguments{
\item{object}{An object of class \code{\link{ScMaSigPro}}.}

\item{value}{The value of the Predictors Matrix.}
}
\value{
Modified `ScMaSigPro` object with new Predictor Matrix.
}
\description{
Replacement method for predictors.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
