% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{predictors}
\alias{predictors}
\title{Get or set the Predictor Matrix}
\usage{
predictors(object, value = "missing")
}
\arguments{
\item{object}{An object of class \code{\link{ScMaSigPro}}.}

\item{value}{The value of the Predictor Matrix.}
}
\value{
Get the Predictor Matrix.
}
\description{
Get or set the Predictor Matrix
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
