% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCoeff.R
\name{queryCoeff}
\alias{queryCoeff}
\title{Get Features from scMaSigPro Object}
\usage{
queryCoeff(
  scmpObj,
  rsq = 0.7,
  p_value = scmpObj@Parameters@p_value,
  includeInflu = TRUE,
  query = "pseudotime_path",
  change = NULL,
  strictly = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{scmpObj}{An object of class \code{\link{ScMaSigPro}}.}

\item{rsq}{Coefficient of determination or R-squared value threshold.}

\item{p_value}{Overall model significance.}

\item{includeInflu}{Whether to include genes with influential observations.}

\item{query}{A string specifying the type of query.}

\item{change}{The effective change at the end of the pseudotime.
(Default is NULL)}

\item{strictly}{If `change != NULL`, selects features if all coefficients
have similar change.}

\item{verbose}{Print detailed output in the console. (Default is TRUE)}
}
\value{
A subset of matrix of coefficients.
}
\description{
This function extracts features from an object of class `scMaSigPro`
based on a specified query. The function query the coefficients matrix.
}
\seealso{
`scMaSigPro::sc.filter()`
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}.
}
