% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.cluster.features.R
\name{sc.cluster.features}
\alias{sc.cluster.features}
\title{Cluster Features in Single-Cell Data}
\usage{
sc.cluster.features(
  scmpObj,
  includeInflu = FALSE,
  cluster.method = "hclust",
  distance = "cor",
  k = 9,
  k.mclust = FALSE,
  cor.methods_covariance = "pairwise.complete.obs",
  cor.method = "pearson",
  hclust.agglo_method = "ward.D"
)
}
\arguments{
\item{scmpObj}{An object of class `scMaSigProClass`. The function checks for
validity and necessary data within this object before proceeding.}

\item{includeInflu}{A boolean indicating whether to include influential points in
the analysis.}

\item{cluster.method}{A string specifying the clustering method to be used.
Options are `"hclust"` for hierarchical clustering, `"kmeans"` for k-means
clustering, and `"mclust"` for model-based clustering.}

\item{distance}{A string specifying the distance measure to be used in
hierarchical clustering. Default is `"cor"` for correlation-based distance.}

\item{k}{The number of clusters to use in k-means or hierarchical clustering.}

\item{k.mclust}{A boolean indicating whether to use k-means clustering for
model-based clustering. Default is `FALSE`.}

\item{cor.methods_covariance}{A string specifying the method for handling missing
values in correlation calculations. Default is `"pairwise.complete.obs"`.}

\item{cor.method}{A string specifying the correlation method to be used.
Default is `"pearson"`.}

\item{hclust.agglo_method}{A string specifying the agglomeration method to be
used in hierarchical clustering. Default is `"ward.D"`.}
}
\value{
The modified `scmpObj` with added clustering information in the
  `@sig.genes@feature.clusters` slot.
}
\description{
Performs clustering on features in single-cell data using hierarchical clustering,
k-means, or model-based clustering methods. It operates on an object of class
`scMaSigProClass` and adds clustering information to this object.
}
