% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.filter.R
\name{sc.filter}
\alias{sc.filter}
\title{Extract significant genes based on R-Square and P-values}
\usage{
sc.filter(
  scmpObj,
  rsq = 0.7,
  p_value = scmpObj@Parameters@p_value,
  vars = c("all", "each", "groups"),
  intercept = "dummy",
  term_p_value = 0.05,
  includeInflu = TRUE
)
}
\arguments{
\item{scmpObj}{An object of class \code{\link{ScMaSigPro}}.}

\item{rsq}{Coefficient of determination or R-squared value threshold.}

\item{p_value}{Overall model significance.}

\item{vars}{Variables for which to extract significant genes. See details.}

\item{intercept}{Specify the branching path treated as reference. See details.
(When `vars` equals "groups").}

\item{term_p_value}{Term wise significance.}

\item{includeInflu}{Whether to include genes with influential observations.}
}
\value{
An object of class \code{\link{ScMaSigPro}}, with updated `Significant`
slot.
}
\description{
`sc.filter()` creates lists of significant genes based on user-specified
constraints.
}
\details{
`vars` Parameter can take one of the following values:
\itemize{
  \item \code{"all"}: Generates one gene list with all significant genes.
  \item \code{"each"}: Generates gene list for each term in the polynomial GLM.
  \item \code{"groups"}: Generates gene list for each branching path.
}

`intercept` Parameter modulates the treatment for intercept coefficients to
apply for selecting significant genes when `vars` equals "groups". There are
three possible values:
 \itemize{
  \item \code{"none"}: No significant intercept (differences) are considered.
  \item \code{"dummy"}: Includes genes with significant intercept differences
  between branching paths.
  \item \code{"all"}: When both significant intercept coefficient for the
  reference path and significant intercept differences are considered for
   selecting significant genes.
}
}
\references{
{Conesa, A., Nueda M.J., Alberto Ferrer, A., Talon, T. 2006.
maSigPro: a Method to Identify Significant Differential Expression Profiles
in Time-Course Microarray Experiments. Bioinformatics 22, 1096-1102}
}
\seealso{
`maSigPro::get.siggenes()`
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}, Ana Conesa and
Maria Jose Nueda, \email{mj.nueda@ua.es}
}
