% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.get.features.R
\name{sc.get.features}
\alias{sc.get.features}
\title{Get Features from scmp Object}
\usage{
sc.get.features(
  scmpObj,
  query = "unique",
  unique.group = "Path1",
  union.ref = "Path1",
  union.target = "Path2vsPath1",
  rsq = 0.7,
  p_value = scmpObj@Parameters@p_value,
  vars = "groups",
  significant.intercept = "dummy",
  term_p_value = 0.05,
  includeInflu = TRUE,
  unique.trend = "any",
  union.ref.trend = "any",
  union.target.trend = "any"
)
}
\arguments{
\item{scmpObj}{An object of class 'scmp'.}

\item{query}{A string specifying the type of query.
Valid options are 'unique' or 'union'.
Default is 'unique'.}

\item{unique.group}{A string specifying the group to consider.
Default is 'Path1'.}

\item{union.ref}{A string specifying the reference group for 'union' query.}

\item{union.target}{A string specifying the target group for 'union' query.}

\item{rsq}{A numeric value specifying the R-squared threshold.}

\item{p_value}{A numeric value, derived from the scmpObj object, for analysis.}

\item{vars}{A string specifying the variables to be used.}

\item{significant.intercept}{A string indicating the treatment of intercepts in the analysis.}

\item{term_p_value}{A numeric value specifying the threshold for term significance.}

\item{includeInflu}{A boolean indicating whether to include influencers in the analysis.}

\item{unique.trend}{A string specifying the trend type for 'unique' query.}

\item{union.ref.trend}{A string specifying the trend type for the reference group in 'union' query.}

\item{union.target.trend}{A string specifying the trend type for the target group in 'union' query.}
}
\value{
Depending on the query, the function returns either a vector of
        unique features (when query is 'unique') or a vector of features
        common to all groups (when query is 'union').
}
\description{
This function extracts features from an object of class 'scmp'
based on a specified query and group. The function supports two types of
queries: 'unique' or 'union', and allows for the selection of a specific
group from the available groups in the object.
}
