% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.get.siggenes.R
\name{sc.get.siggenes}
\alias{sc.get.siggenes}
\title{Extract significant genes for sets of variables in time series gene expression experiments}
\usage{
sc.get.siggenes(
  scmpObj,
  rsq = 0.7,
  add.IDs = FALSE,
  IDs = NULL,
  matchID.col = 1,
  only.names = FALSE,
  vars = c("all", "each", "groups"),
  significant.intercept = "dummy",
  groups.vector = NULL,
  trat.repl.spots = "none",
  index = IDs[, (matchID.col + 1)],
  match = IDs[, matchID.col],
  r = 0.7
)
}
\arguments{
\item{scmpObj}{A \code{T.fit} object.}

\item{rsq}{Cut-off level at the R-squared value for the stepwise regression fit. Only genes with R-squared more than 'rsq' are selected.}

\item{add.IDs}{Logical indicating whether to include additional gene id's in the result.}

\item{IDs}{Matrix containing additional gene id information (required when \code{add.IDs = TRUE}).}

\item{matchID.col}{Number of the matching column in the matrix \code{IDs} for adding gene ids.}

\item{only.names}{Logical. If \code{TRUE}, expression values are omitted in the results.}

\item{vars}{Variables for which to extract significant genes.}

\item{significant.intercept}{Experimental groups for which significant intercept coefficients are considered.}

\item{groups.vector}{Required when \code{vars = "groups"}.}

\item{trat.repl.spots}{Treatment given to replicate spots. Possible values are \code{"none"} and \code{"average"}.}

\item{index}{Argument of the \code{\link{average.rows}} function to use when \code{trat.repl.spots = "average"}.}

\item{match}{Argument of the \code{\link{average.rows}} function to use when \code{trat.repl.spots = "average"}.}

\item{r}{Minimum Pearson correlation coefficient for replicated spots profiles to be averaged.}
}
\value{
A list with two elements:
\itemize{
  \item \code{summary}: A vector or matrix listing significant genes for the variables given by the function parameters.
  \item \code{sig.genes}: A list with detailed information on the significant genes found for the variables given by the function parameters.
  Each element of the list is also a list containing:
    \describe{
      \item{\code{sig.profiles}:}{Expression values of significant genes.}
      \item{\code{coefficients}:}{Regression coefficients of the adjusted models.}
      \item{\code{group.coeffs}:}{Regression coefficients of the implicit models of each experimental group.}
      \item{\code{sig.pvalues}:}{P-values of the regression coefficients for significant genes.}
      \item{\code{g}:}{Number of genes.}
      \item{\code{...}:}{Arguments passed by previous functions.}
    }
  }
}
\description{
There are 3 possible values for the \code{vars} argument:
\itemize{
  \item \code{"all"}: generates one single matrix or gene list with all significant genes.
  \item \code{"each"}: generates as many significant genes extractions as variables in the general regression model.
  \item \code{"groups"}: generates a significant genes extraction for each experimental group.
}
}
\details{
Extract significant genes for sets of variables in time series gene expression experiments

This function creates lists of significant genes for a set of variables whose significance
value has been computed with the \code{T.fit} function.


Refer to the function description for details on the arguments and their usage.
}
\examples{
# Example usage of the function can be placed here.

}
\references{
Conesa, A., Nueda M.J., Alberto Ferrer, A., Talon, T. (2006).
maSigPro: a Method to Identify Significant Differential Expression Profiles in Time-Course Microarray Experiments.
Bioinformatics, 22(9), 1096-1102. \url{https://doi.org/10.1093/bioinformatics/btl056}
}
\author{
Ana Conesa and Maria Jose Nueda (mj.nueda@ua.es)
}
\keyword{manip}
