% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.make.design.matrix.R
\name{sc.make.design.matrix}
\alias{sc.make.design.matrix}
\title{Create design matrix for 'scMaSigProClass' object}
\usage{
sc.make.design.matrix(scmpObj, degree = 2, time.col, path.col)
}
\arguments{
\item{scmpObj}{A 'scMaSigProClass' object.}

\item{degree}{Degree of the design matrix (default: 2).}

\item{time.col}{Name of the time column.}

\item{path.col}{Name of the path column.}
}
\value{
Returns the 'scmpObj' with an updated 'edesign' slot.
}
\description{
This function creates a design matrix using the 'compress.sce' slot of a 'scMaSigProClass' object.
It generates an 'edesignClass' object which is then stored in the 'edesign' slot of the 'scMaSigProClass' object.
}
\examples{
# Insert an example of how to use the function here.

}
