% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.plot.bins.R
\name{sc.plot.bins}
\alias{sc.plot.bins}
\title{sc.plot.bins: Plot Bin Sizes Across Binned Time and Paths}
\usage{
sc.plot.bins(
  scmpObj,
  add_tile = T,
  bar.title = "Bin size across binned pseudotime",
  tile.title = "Bin size across paths along binned pseudotime"
)
}
\arguments{
\item{scmpObj}{A SingleCellExperiment object with an additional slot '@compress.sce' that contains compression information.}

\item{add_tile}{A logical flag indicating whether to add a tile (heatmap) plot alongside the bar plot. Default is TRUE.}

\item{bar.title}{A title of the barplot}

\item{tile.title}{A title of the heatmap}
}
\value{
A bar plot and, optionally, a tile (heatmap) plot, visualizing the bin sizes across different binned time and paths.
If add_tile is TRUE, returns a combined ggplot object with both plots; otherwise, only the bar plot is printed.
}
\description{
This function generates plots to visualize the compressed data from a SingleCellExperiment object.
It produces a bar plot and, optionally, a tile (heatmap) plot to display the bin sizes across different
binned time intervals and paths.
}
