% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.PlotProfiles.R
\name{sc.PlotProfiles}
\alias{sc.PlotProfiles}
\title{Plot Groups Function}
\usage{
sc.PlotProfiles(
  scmpObj,
  xlab = "Pooled Pseudotime",
  ylab = "Pseudobulk Expression",
  groupBy = "coeff",
  logs = TRUE,
  logType = "log",
  smoothness = 0.01,
  includeInflu = TRUE
)
}
\arguments{
\item{scmpObj}{object of class scmpObj}

\item{xlab}{X-axis label. Default is "Pooled Pseudotime".}

\item{ylab}{Y-axis label. Default is "Pseudobulk Expression".}

\item{groupBy}{A vector of maximum length 6 or a single feature of ID}

\item{logs}{Whether to plot log of counts}

\item{logType}{Log type required}

\item{smoothness}{description}

\item{includeInflu}{description}
}
\value{
Generates a plot.
}
\description{
This function generates plots based on various parameters.
}
