% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.set.poly.R
\name{sc.set.poly}
\alias{sc.set.poly}
\title{Set up polynomial models and create Predictor Matrix}
\usage{
sc.set.poly(
  scmpObj,
  poly_degree = 2,
  bin_ptime_col = scmpObj@Parameters@bin_ptime_col,
  path_col = scmpObj@Parameters@path_col
)
}
\arguments{
\item{scmpObj}{An object of class \code{\link{ScMaSigPro}}.}

\item{poly_degree}{Degree of the polynomial.}

\item{bin_ptime_col}{A character string representing the column name
for binned Pseudotime values in 'Dense' data.}

\item{path_col}{A character string representing the column name for branching
path assignment in 'Sparse' or 'Dense' slot.}
}
\value{
An object of class \code{\link{ScMaSigPro}}, with updated `Design`
slot.
}
\description{
Set up polynomial models and create Predictor Matrix that will contain the
independent variables. It is a wrapper around `maSigPro::make.design.matrix`.
}
\references{
{Conesa, A., Nueda M.J., Alberto Ferrer, A., Talon, T. 2006.
maSigPro: a Method to Identify Significant Differential Expression Profiles
in Time-Course Microarray Experiments. Bioinformatics 22, 1096-1102}
}
\seealso{
\code{\link{MatrixDesign}} Class.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}, Ana Conesa and
Maria Jose Nueda, \email{mj.nueda@ua.es}
}
