% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.t.fit.R
\name{sc.t.fit}
\alias{sc.t.fit}
\title{Perform stepwise regression fit to select for significant terms.}
\usage{
sc.t.fit(
  scmpObj,
  selection_method = "backward",
  p_value = scmpObj@Parameters@p_value,
  nvar_correction = FALSE,
  family = scmpObj@Parameters@distribution,
  epsilon = scmpObj@Parameters@epsilon,
  offset = scmpObj@Parameters@offset,
  verbose = TRUE,
  parallel = FALSE,
  log_offset = scmpObj@Parameters@log_offset,
  max_it = scmpObj@Parameters@max_it
)
}
\arguments{
\item{scmpObj}{An object of class \code{\link{ScMaSigPro}}.}

\item{selection_method}{Method for step-wise regression.}

\item{p_value}{Significance level used for variable selection in the stepwise
regression.}

\item{nvar_correction}{Argument for correcting significance level. See details.}

\item{family}{Distribution of the error term.}

\item{epsilon}{Model convergence tolerance.}

\item{offset}{A logical value specifying whether to use offset during fitting.}

\item{verbose}{Print detailed output in the console. (Default is TRUE)}

\item{parallel}{Use forking process to run parallelly. (Default is FALSE)
(Currently, Windows is not supported)}

\item{log_offset}{A logical value specifying whether to take the logarithm of
the offsets.}

\item{max_it}{Maximum number of iterations to fit the model.}
}
\value{
An object of class \code{\link{ScMaSigPro}}, with updated `Estimate`
slot.
}
\description{
Performs stepwise regression and selects the significant polynomial terms
from the full polynomial model. This function is succeeded by
\code{scMaSigPro::sc.p.vector()}.
}
\references{
{Conesa, A., Nueda M.J., Alberto Ferrer, A., Talon, T. 2006.
maSigPro: a Method to Identify Significant Differential Expression Profiles
in Time-Course Microarray Experiments. Bioinformatics 22, 1096-1102}
}
\seealso{
\code{\link{Estimates}} Class.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}, Ana Conesa and
Maria Jose Nueda, \email{mj.nueda@ua.es}
}
\keyword{models}
\keyword{regression}
