% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maSigPro_term_selection.R
\name{sc.two.ways.stepback}
\alias{sc.two.ways.stepback}
\title{Polynomial term selection Methods from maSigPro}
\usage{
sc.two.ways.stepback(
  y = y,
  d = d,
  alfa = 0.05,
  family = gaussian(),
  epsilon = 1e-05,
  useOffset,
  useWeight,
  max_it
)
}
\description{
Polynomial term selection Methods from maSigPro
}
\references{
{Conesa, A., Nueda M.J., Alberto Ferrer, A., Talon, T. 2006.
maSigPro: a Method to Identify Significant Differential Expression Profiles in Time-Course Microarray Experiments.
Bioinformatics 22, 1096-1102}
}
\author{
{Ana Conesa and Maria Jose Nueda, \email{mj.nueda@ua.es}}
}
\keyword{internal}
