% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_makeDesign.R
\name{sc_makeDesign}
\alias{sc_makeDesign}
\title{Create Object for Sc MaSigPro}
\usage{
sc_makeDesign(
  sce.object,
  path.col = "Group",
  time.col = "Step",
  poly.order = 2
)
}
\arguments{
\item{sce.object}{Object of the \link[SingleCellExperiment]{SingleCellExperiment} class.}

\item{path.col}{Column name having path information.}

\item{time.col}{Column name having pseudotime information.}

\item{poly.order}{Max degree of the polynomial}
}
\description{
Create Object for Sc MaSigPro
}
