% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_pVector.R
\name{sc_pVector}
\alias{sc_pVector}
\title{sc_pVector}
\usage{
sc_pVector(
  scMaSigPro.obj,
  assay.name = "counts",
  p.vector.sig = 0.05,
  m.t.c.method = "BH",
  min.counts = 6,
  model.type = "nb",
  model.control = list(ep = 1e-05, nb.k = 10, link = "log"),
  use.offset = FALSE
)
}
\arguments{
\item{scMaSigPro.obj}{An object of the class scMaSigProClass should be supplied.}

\item{assay.name}{The name of the assay utilized from the Single Cell experiment class.}

\item{p.vector.sig}{The significance level of the p-value for full vs intercept models.}

\item{m.t.c.method}{Method for multiple testing correction of p.value with  \code{\link[stats]{p.adjust}}}

\item{min.counts}{Genes with less than this number of true numerical values will be excluded from the analysis.}

\item{model.type}{One of the  "g": Gaussian, "nb", "zinb": Zero-Inflated Negative Binomial,
"p": Poisson, "zip": Zero-Inflated Poisson, "zap": Zero-Altered Poisson, "zanb": Zero-Altered Negative Binomial,
"gp": Generalized Poisson, "gnb": Generalized Negative Binomial'.}

\item{model.control}{A list of control parameters for the model estimation. For e.g.
 The available control parameters for 'nb': Negative Binomial are:
\describe{
     \item{\code{ep}}{The epsilon value for model convergence. Default is 0.00001.}
     \item{\code{nb.k}}{The size parameter for the negative binomial distribution. Default is 10.}
  }}

\item{use.offset}{If set to true then log(Size factors) will be used as offsets}
}
\description{
sc_pVector
}
