% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_tFit.R
\name{sc_tFit}
\alias{sc_tFit}
\title{sc_tFit}
\usage{
sc_tFit(
  scMaSigPro.obj,
  covar.sig = 0.05,
  min.counts = 6,
  model.type = "nb",
  model.control = list(ep = 1e-05, nb.k = 10, link = "log")
)
}
\arguments{
\item{scMaSigPro.obj}{An object of the class scMaSigProClass should be supplied.}

\item{covar.sig}{The significance level of the p-value for each covariate during model selection.}

\item{min.counts}{Genes with less than this number of true numerical values will be excluded from the analysis.}

\item{model.type}{One of the  "g": Gaussian, "nb", "zinb": Zero-Inflated Negative Binomial,
"p": Poisson, "zip": Zero-Inflated Poisson, "zap": Zero-Altered Poisson, "zanb": Zero-Altered Negative Binomial,
"gp": Generalized Poisson, "gnb": Generalized Negative Binomial'.}

\item{model.control}{A list of control parameters for the model estimation. For e.g.
 The available control parameters for 'nb': Negative Binomial are:
\describe{
     \item{\code{ep}}{The epsilon value for model convergence. Default is 0.00001.}
     \item{\code{nb.k}}{The size parameter for the negative binomial distribution. Default is 10.}
  }}
}
\description{
sc_tFit
}
