% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlyDocs_sce.test.R
\docType{data}
\name{sce.test}
\alias{sce.test}
\title{Title of the dataset}
\format{
A brief description of the format of the dataset.
}
\source{
Where the dataset comes from.
}
\usage{
sce.test
}
\description{
A brief description about the dataset.
}
\references{
Any references associated with the dataset.
}
\keyword{datasets}
