% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.sel.features.R
\name{scMaSigPro_t_fit}
\alias{scMaSigPro_t_fit}
\alias{sc.sel.features}
\title{pvector_scMASigPro}
\usage{
sc.sel.features(
  scMaSigPro.obj,
  Q = 0.05,
  filter = "rsq",
  dist.family = "nb",
  assay.name = "counts",
  ep = 1e-05,
  method = "stepback",
  nb.k = 1
)
}
\arguments{
\item{scMaSigPro.obj}{Description of Slot}

\item{Q}{Description of Slot}

\item{dist.family}{Description of Slot}

\item{assay.name}{Description of Slot}

\item{ep}{Epsilon}

\item{method}{Step-bacl}

\item{nb.k}{Dispersion factor negative binomial}

\item{covar.Q}{description}
}
\description{
pvector_scMASigPro
}
