% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpClass.R
\docType{class}
\name{scMaSigProClass-class}
\alias{scMaSigProClass-class}
\title{scMaSigProClass}
\description{
A class to represent the ScMaSigPro analysis results and associated data.
Inherits from \code{SingleCellExperiment}.
}
