% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpClass.R
\docType{class}
\name{scMaSigProClass}
\alias{scMaSigProClass}
\title{scMaSigProClass}
\description{
A class to represent the ScMaSigPro analysis results and associated data.
Inherits from \code{SingleCellExperiment}.
}
\section{Slots}{

\describe{
\item{\code{covariate}}{A \code{covariateClass} object containing covariate information.}

\item{\code{parameters}}{A \code{parameterClass} object containing analysis parameters.}

\item{\code{pVector}}{A \code{pVectorClass} object containing p-vector information.}

\item{\code{tFit}}{A \code{tFitClass} object containing model fitting information.}

\item{\code{sigGenes}}{A \code{sigGeneClass} object containing significant genes information.}
}}

