% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpClass.R
\docType{class}
\name{scmp}
\alias{scmp}
\alias{scmp-class}
\title{scmp}
\description{
A class to represent the ScMaSigPro analysis results and associated data.
Inherits from \code{SingleCellExperiment}.
}
\section{Slots}{

\describe{
\item{\code{sparse}}{Object of Class SingleCellExperiment. See \pkg{SingleCellExperiment} for more details.}

\item{\code{profile}}{Object of Class sigProfileClass See \pkg{sigProfileClass} for more details.}

\item{\code{estimate}}{Object of Class estimateClass See \code{\link{estimateClass}} for more details.}

\item{\code{dense}}{Object of Class SingleCellExperiment. See \pkg{SingleCellExperiment} for more details.}

\item{\code{design}}{Object of Class designClass. See \code{\link{designClass}} for more details.}

\item{\code{param}}{Object of Class paramClass. See \code{\link{paramClass}} for more details.}

\item{\code{sig.genes}}{ABC}
}}

\keyword{classes}
