% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{scmp.ob}
\alias{scmp.ob}
\title{scMaSigPro Object with results}
\format{
An object of class \code{ScMaSigPro} of length 1.
}
\usage{
# Loading
data("scmp.ob", package = "scMaSigPro")
}
\description{
A small scMaSigPro object created from simulated data. Please follow Vignette
`scMaSigPro: Quick Start Guide` for more details.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
\keyword{datasets}
