% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{scmp_write}
\alias{scmp_write}
\title{Write scMaSigPro object to file.}
\usage{
scmp_write(
  scmpObj,
  filename,
  filepath = NULL,
  format = "rds",
  overwrite = FALSE
)
}
\arguments{
\item{filename}{filename. Please donot provide extensions.}

\item{filepath}{File path.}

\item{format}{File format. Default is 'rds'.}

\item{overwrite}{Overwrite existing file. Default is FALSE.}

\item{object}{scMaSigPro object.}

\item{verbose}{Print verbose output. Default is FALSE.}
}
\value{
Write a scMaSigPro object to file.
}
\description{
Write scMaSigPro object to various file formats.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}.
}
