% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_io.R, R/scmp_write_h5.R
\name{scmp_write_h5}
\alias{scmp_write_h5}
\title{Write scMaSigPro object to file.}
\usage{
scmp_write_h5(scmpObj, file_path, overwrite = FALSE, verbose = TRUE)

scmp_write_h5(scmpObj, file_path, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{overwrite}{Overwrite existing file. Default is FALSE.}

\item{verbose}{Print verbose output. Default is FALSE.}

\item{object}{scMaSigPro object.}

\item{filename}{filename. Please donot provide extensions.}

\item{format}{File format. Default is 'rds'.}

\item{scmpObject}{scMaSigPro object.}

\item{filepath}{File path.}
}
\value{
Write a scMaSigPro object to file.

Write a scMaSigPro object to H5 file.
}
\description{
Write scMaSigPro object to various file formats.

Write scMaSigPro object H5 file.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}.
}
\keyword{internal}
