% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.p.vector.class.R
\docType{class}
\name{scPVectorClass}
\alias{scPVectorClass}
\alias{scPVectorClass-class}
\title{A class to store regression fit results for time series gene expression experiments}
\description{
The scPVectorClass is designed to hold the results of a regression fit for each gene in a time series gene expression experiment.
It contains slots to store various results, such as expression values for significant genes, computed p-values, adjusted p-values,
number of input genes, number of genes taken in the regression fit, and more. This class is useful for analyzing time course
microarray experiments and identifying significant differential expression profiles.
}
\section{Slots}{

\describe{
\item{\code{SELEC}}{dgCMatrix containing the expression values for significant genes.}

\item{\code{p.vector}}{Numeric vector containing the computed p-values.}

\item{\code{p.adjusted}}{Numeric vector of FDR-adjusted p-values.}

\item{\code{FDR}}{P-value at FDR \code{Q} control when Benjamini & Hochberg (BH) correction is used.}

\item{\code{dis}}{Data frame containing the matrix used in the regression fit.}

\item{\code{groups.vector}}{Character list containing groups information.}

\item{\code{family}}{Distribution function to be used in the glm model. If NULL, the
family will be \code{negative.binomial(theta)} when \code{counts = TRUE} or
\code{gaussian()} when \code{counts = FALSE}.}
}}

\examples{
#### GENERATE TIME COURSE DATA
## generates n random gene expression profiles of a data set with
## one control plus 3 treatments, 3 time points and r replicates per time point.

# ... (add example usage of the function here)

}
\seealso{
\code{\link{T.fit}}, \code{\link{lm}}
}
\author{
Priyansh Srivastava <spriyansh29@gmail.com>
}
\keyword{classes}
\keyword{regression}
