% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.T.fit.class.R
\docType{class}
\name{scTFitClass}
\alias{scTFitClass}
\alias{scTFitClass-class}
\title{Class "scTFitClass"}
\description{
A class for fitting a model.
}
\section{Slots}{

\describe{
\item{\code{sol}}{A data frame representing ...}

\item{\code{sig.profiles}}{A data frame representing ...}

\item{\code{coefficients}}{A data frame representing ...}

\item{\code{group.coeffs}}{A data frame representing ...}

\item{\code{t.score}}{A data frame representing ...}

\item{\code{variables}}{A character vector representing ...}

\item{\code{G}}{An integer representing ...}

\item{\code{g}}{An integer representing ...}

\item{\code{dat}}{A matrix representing ...}

\item{\code{dis}}{A data frame representing ...}

\item{\code{step.method}}{A character representing ...}

\item{\code{groups.vector}}{A character vector representing ...}

\item{\code{edesign}}{A matrix representing ...}

\item{\code{influ.info}}{A matrix representing ...}
}}

\keyword{classes}
