% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectShortPath.R
\name{select_longer_vector}
\alias{select_longer_vector}
\title{Select the Longer of Two Vectors}
\usage{
select_longer_vector(vector1, vector2, vector1_label, vector2_label)
}
\arguments{
\item{vector1}{A numeric vector.}

\item{vector2}{A numeric vector.}

\item{vector1_label}{A label (character string) for `vector1`.}

\item{vector2_label}{A label (character string) for `vector2`.}
}
\value{
A list containing:
\itemize{
  \item \code{long_vec}: The longer vector of `vector1` and `vector2`.
  \item \code{long_vec_label}: The label corresponding to the longer vector.
  \item \code{short_vec}: The shorter vector of `vector1` and `vector2`.
  \item \code{short_vec_label}: The label corresponding to the shorter vector.
}
If both vectors are of the same length, the function returns 1.
}
\description{
This function compares the lengths of two vectors and returns the longer one.
If both vectors have the same length, the function returns 1.
}
\keyword{internal}
