% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.selectPath.R, R/selectPath.m3.R
\name{selectPath.m3}
\alias{selectPath.m3}
\title{Subset a CDS object interactively with Shiny}
\usage{
selectPath.m3(
  cdsObj,
  redDim = "umap",
  annotation = "cell.type",
  returnType = "scmpObj"
)

selectPath.m3(
  cdsObj,
  redDim = "umap",
  annotation = "cell.type",
  returnType = "scmpObj"
)
}
\arguments{
\item{cdsObj}{An cdsObject of class `scMaSigProClass`. This cdsObject will be checked
to ensure it's the right type.}

\item{redDim}{Dimension to use for the plot}

\item{annotation}{A character vector indicating the paths to be selected.}

\item{returnType}{'scmpObj' or subseted 'cds'}

\item{annotation_col}{A character vector indicating the paths to be selected.}

\item{pseudotime_col}{Name of the column with Pseudotime}

\item{path_col}{Name of the column with Path}
}
\value{
A `scMaSigProClass` object, subsetted based on the specified paths.

A `scMaSigProClass` object, subsetted based on the specified paths.
}
\description{
Subset a CDS object interactively with Shiny

Subset a CDS object interactively with Shiny
}
