% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPath.R
\name{selectPath}
\alias{selectPath}
\title{Select Paths for scMaSigPro}
\usage{
selectPath(
  obj,
  sel.path,
  balance_paths = T,
  pathCol = "Path",
  plot_paths = T,
  pTimeCol = "Pseudotime",
  verbose = T
)
}
\arguments{
\item{obj}{An object of class `scMaSigProClass`. This object will be checked
to ensure it's the right type.}

\item{sel.path}{A character vector indicating the paths to be selected.}

\item{balance_paths}{Logical indicating if paths should be balanced. Default is TRUE.}

\item{pathCol}{The column name in `colData(sceObj)` that indicates the path.
Defaults to "Path".}

\item{plot_paths}{Logical indicating if paths should be plotted. Default is TRUE.}

\item{pTimeCol}{The column name indicating the pseudotime. Default is "Pseudotime".}

\item{verbose}{Logical indicating if verbose messages should be displayed. Default is TRUE.}
}
\value{
A `SingleCellExperiment` object subsetted based on the specified paths.
}
\description{
This function selects paths to be used in scMaSigPro. It extracts
a sub-object based on the specified paths.
}
\examples{
# Assuming you have an example object of class `scMaSigProClass`:
# selected_obj <- selectPath(example_obj, sel.path = c("Path1", "Path3"))

}
