% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySelect.R
\name{shinySelect}
\alias{shinySelect}
\title{Function to interactively select nodes from shiny}
\usage{
shinySelect(
  trajectory_data,
  annotation_data,
  label_coords,
  ptime_col,
  inputType = "Monocle3"
)
}
\arguments{
\item{trajectory_data}{Dataframe with trajectory data.}

\item{annotation_data}{Dataframe for with annotation data.}

\item{label_coords}{How to label corrdinates.}

\item{inputType}{Name of the method.}
}
\description{
Function to interactively select nodes from shiny
}
\keyword{internal}
