% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_functions.R
\name{showCoeff}
\alias{showCoeff}
\title{Show or Return the Coefficent matrix}
\usage{
showCoeff(scmpObj, view = FALSE, return = TRUE, includeInflu = TRUE)
}
\arguments{
\item{scmpObj}{An object of class \code{\link{ScMaSigPro}}.}

\item{view}{Whether to view the data in the explorer. (Default: FALSE)}

\item{return}{Whether to return the data. (Default: TRUE)}

\item{includeInflu}{Whether to include genes with inluential observations.}
}
\value{
The computed Coefficent matrix as a dataframe.
}
\description{
This function is used to view or return the coefficents from the provided
scMaSigPro object.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
