% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateShow.R
\name{showEstimates}
\alias{showEstimates}
\title{Show Estimated Coefficients Function}
\usage{
showEstimates(scMaSigPro.obj, view = TRUE)
}
\arguments{
\item{scMaSigPro.obj}{An object of class "scMaSigPro" representing the output
of the `scMaSigPro` function.}

\item{view}{Logical value indicating whether to view the matrix of beta
estimates in the RStudio Viewer pane. Default is TRUE.}
}
\value{
A transposed matrix containing the estimated coefficients (beta
        estimates) from the scMaSigPro object.
}
\description{
This function extracts the estimated coefficients (beta estimates) from the
scMaSigPro object and returns them as a transposed matrix. The function
provides the option to view the matrix using the "View" function.
}
\examples{
# Assuming scMaSigPro.obj is an object returned from the scMaSigPro function
# View the estimated coefficients
showEstimates(scMaSigPro.obj)

# Get the estimated coefficients without viewing
estimates <- showEstimates(scMaSigPro.obj, view = FALSE)

}
\keyword{internal}
