% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_functions.R
\name{showGroupCoeff}
\alias{showGroupCoeff}
\title{Show or Return the Branching Path Coefficent matrix}
\usage{
showGroupCoeff(scmpObj, view = FALSE, return = TRUE, includeInflu = TRUE)
}
\arguments{
\item{scmpObj}{An object of class \code{\link{ScMaSigPro}}.}

\item{view}{Whether to view the data in the explorer. (Default: FALSE)}

\item{return}{Whether to return the data. (Default: TRUE)}

\item{includeInflu}{Whether to include genes with inluential observations.}
}
\value{
The computed branching path coefficent matrix as a dataframe.
}
\description{
This function is used to view or return the branching paths coefficents from
the provided scMaSigPro object.
}
