% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_functions.R
\name{showPoly}
\alias{showPoly}
\title{Print the full model formula.}
\usage{
showPoly(scmpObj)
}
\arguments{
\item{scmpObj}{An object of class \code{\link{ScMaSigPro}}.}
}
\value{
Character string of the formula for the full model.
}
\description{
Print the full model formula in console as a string.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
