% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_functions.R
\name{showSigProf}
\alias{showSigProf}
\title{Show or Return the counts for non-flat profile.}
\usage{
showSigProf(scmpObj, view = FALSE, return = TRUE, includeInflu = FALSE)
}
\arguments{
\item{scmpObj}{An object of class \code{\link{ScMaSigPro}}.}

\item{view}{Whether to view the data in the explorer. (Default: FALSE)}

\item{return}{Whether to return the data. (Default: TRUE)}

\item{includeInflu}{Whether to include genes with influential observations.}
}
\value{
Pseudo-bulk counts as matrix for genes with non-flat profiles.
}
\description{
This function is used to view or return the pseudo-bulk counts of the genes
with non-flat profiles. from the provided scMaSigPro object.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
