% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscoreShow.R
\name{showTscores}
\alias{showTscores}
\title{Show T-scores Matrix}
\usage{
showTscores(scMaSigPro.obj, view = TRUE)
}
\arguments{
\item{scMaSigPro.obj}{A \code{scMaSigPro_t_fit} object containing the results of the ScMaSigPro analysis.}

\item{view}{Logical value indicating whether to view the T-scores matrix using a data viewer (e.g., RStudio Viewer).}
}
\value{
The T-scores matrix extracted from the ScMaSigPro analysis results.
}
\description{
This function extracts and displays the T-scores matrix from the output of the ScMaSigPro
analysis (\code{scMaSigPro_t_fit} object).
}
\keyword{internal}
