% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpClass.R
\docType{class}
\name{sigClass}
\alias{sigClass}
\alias{sigClass-class}
\title{sigClass}
\description{
sigClass
}
\section{Slots}{

\describe{
\item{\code{sig.genes}}{list}

\item{\code{feature.clusters}}{list}
}}

\keyword{classes}
