% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpClass.R
\docType{class}
\name{sigGeneClass}
\alias{sigGeneClass}
\title{sigGeneClass}
\description{
A class to represent significant genes and their associations.
}
\section{Slots}{

\describe{
\item{\code{group.associations}}{A list representing group associations.}

\item{\code{sel.genes.frame}}{A list representing selected genes frame.}
}}

