% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScMaSigProClass.R
\docType{class}
\name{Significant}
\alias{Significant}
\alias{Significant-class}
\title{Significant}
\description{
S4 class to store significantly selected features and their clusters.
}
\section{Slots}{

\describe{
\item{\code{genes}}{Named list of genes selected based on user-specified thresholds.}

\item{\code{clusters}}{Named list of clusters selected based on user-specified
thresholds.}
}}

\keyword{classes}
