% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpClass.R
\docType{class}
\name{sigProfileClass}
\alias{sigProfileClass}
\alias{sigProfileClass-class}
\title{A class to store regression fit results for time series gene expression experiments}
\description{
The sigProfileClass is designed to hold the results of a regression fit for each gene in a time series gene expression experiment.
It contains slots to store various results, such as expression values for significant genes, computed p-values, adjusted p-values,
number of input genes, number of genes taken in the regression fit, and more. This class is useful for analyzing time course
microarray experiments and identifying significant differential expression profiles.
}
\section{Slots}{

\describe{
\item{\code{non.flat}}{a charcter vector}

\item{\code{p.vector}}{Numeric vector containing the computed p-values.}

\item{\code{p.adjusted}}{Numeric vector of FDR-adjusted p-values.}

\item{\code{FDR}}{P-value at FDR \code{Q} control when Benjamini & Hochberg (BH) correction is used.}
}}

\seealso{
\code{\link{T.fit}}, \code{\link{lm}}
}
\author{
Priyansh Srivastava <spriyansh29@gmail.com>
}
\keyword{classes}
\keyword{regression}
