% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim2Path.R
\docType{data}
\name{sim.sce}
\alias{sim.sce}
\title{Simulated SingleCellExperiment Object created using Splatter}
\format{
An object of class \code{SingleCellExperiment} with 500 rows and 200 columns.
}
\source{
Source of the data, if applicable
}
\usage{
sim.sce
}
\description{
A small simulated SingleCellExperiment Object created using Splatter. The dataset
contains 200 cells and 500 genes. The dataset is simulated to have a bifurcating
topology of the trajectory.
}
\details{
__
}
\references{
References, if applicable
}
\keyword{datasets}
