% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_by_smallest_range.R
\name{slice_by_smallest_range}
\alias{slice_by_smallest_range}
\title{Slice Data By The Smallest Pseudotime Range}
\usage{
slice_by_smallest_range(data, path_col, pseudotime_col, verbose = T)
}
\arguments{
\item{data}{A dataframe containing single cell data with columns specified by `path_col` and `pseudotime_col`.}

\item{path_col}{A string specifying the column name in `data` that contains path identifiers.}

\item{pseudotime_col}{A string specifying the column name in `data` that contains pseudotime values.}

\item{verbose}{A logical value indicating whether to print informational messages. Default is `TRUE`.}
}
\value{
A dataframe where paths have been trimmed to match the pseudotime range of the path with the smallest range.
Cells outside this range for each path are removed.
}
\description{
This function trims data from paths based on the path with the smallest pseudotime range.
Specifically, it determines which path has the smallest range of pseudotimes
and then slices all other paths to match this range.
}
