% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squeeze.R
\name{squeeze}
\alias{squeeze}
\title{Wrapper for scMaSigPro Binning}
\usage{
squeeze(
  scmpObject,
  pseudotime_colname = "Pseudotime",
  path_colname = "Path",
  bin_method = "Sturges",
  drop.fac = 0.5,
  verbose = TRUE,
  cluster_count_by = "sum",
  assay_name = "counts",
  binning = "universal",
  additional_params = NULL,
  bin_pseudotime_colname = "scmp_binned_pseudotime"
)
}
\arguments{
\item{scmpObject}{An object.}

\item{pseudotime_colname}{Name of the column in `cell.metadata` generated using
\code{\link[SingleCellExperiment]{colData}} storing information for Pseudotime.
(Default is "Pseudotime")}

\item{path_colname}{Name of the column in `cell.metadata` generated using
\code{\link[SingleCellExperiment]{colData}} storing information for Path.
(Default is `path_prefix`)}

\item{bin_method}{A character string (default = "Sturges"). The method to be
used to estimate the optimal number of bins.}

\item{drop.fac}{A numeric value (default = 0.5). The factor by which to
decrease the number of bins if the initial binning results in too many bins.}

\item{verbose}{Print detailed output in the console. (Default is TRUE)}

\item{cluster_count_by}{A character string (default = "sum"). The method to
use to aggregate counts within each cluster.}

\item{assay_name}{The name of the assay (default: "counts").}

\item{binning}{A character string (deafult = "universal"). When set to
"individual", the bins are calculated per path iteratively.}

\item{additional_params}{Additional Parameters in a list
The name of the column in the design file that contains the binned data.}
}
\value{
Returns the object of 'scMaSigProClass' with updated slots.

#' @examples
\dontrun{

}
}
\description{
`squeeze()` performs a series of operations including data compression, creating pseudo-bulk-design, and creating pseudo-bulk counts.
It then generates a SingleCellExperiment object with updated slots.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
