% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpClass.R
\docType{class}
\name{tFitClass}
\alias{tFitClass}
\title{tFitClass}
\description{
A class to represent model fitting information.
}
\section{Slots}{

\describe{
\item{\code{covar.models}}{A list representing covariate models.}

\item{\code{model.attributes}}{A list representing model attributes.}
}}

