% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScMaSigProClass.R
\docType{class}
\name{VariableProfiles}
\alias{VariableProfiles}
\alias{VariableProfiles-class}
\title{VariableProfiles}
\description{
S4 class to store results of the global model fitting for all the features.
}
\section{Slots}{

\describe{
\item{\code{non.flat}}{a charcter vector}

\item{\code{p.vector}}{Numeric vector containing the computed p-values.}

\item{\code{p.adjusted}}{Numeric vector of FDR-adjusted p-values.}

\item{\code{FDR}}{P-value at FDR \code{Q} control when Benjamini & Hochberg (BH) correction is used.}
}}

\seealso{
\code{\link{T.fit}}, \code{\link{lm}}
}
\author{
Priyansh Srivastava <spriyansh29@gmail.com>
}
\keyword{classes}
