% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warningCondition.R
\name{warningCondition}
\alias{warningCondition}
\title{Custom Warning Function}
\usage{
warningCondition(condition, message)
}
\arguments{
\item{condition}{A logical value indicating whether the warning should be triggered.}

\item{message}{A character string specifying the warning message.}
}
\description{
This function generates a warning with a custom message if a given condition is TRUE.
}
\examples{
# Generate a warning if a condition is TRUE
warningCondition(TRUE, "This is a custom warning message.")

# Generate no warning if the condition is FALSE
warningCondition(FALSE, "This message won't trigger a warning.")

}
\keyword{internal}
