% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addParamClass.R
\docType{class}
\name{addParamClass}
\alias{addParamClass}
\alias{addParamClass-class}
\title{Class "addParamClass"}
\description{
Class "addParamClass"
}
\section{Slots}{

\describe{
\item{\code{bin_pseudotime_colname}}{A character representing the name of the column for binned pseudotime values.}

\item{\code{path_prefix}}{A character representing the prefix for path labeling.}

\item{\code{root_label}}{A character representing the label for the root of the tree.}

\item{\code{pseudotime_colname}}{A character representing the name of the column for pseudotime values.}

\item{\code{bin_method}}{A character representing the algorithm used for binning.}

\item{\code{path_colname}}{A character representing the name of the column for path values.}

\item{\code{bin_colname}}{A character representing the name of the column for bin values.}

\item{\code{bin_size_colname}}{A character representing the name of the column for bin sizes.}

\item{\code{bin_members_colname}}{A character representing the name of the column for bin members.}

\item{\code{Q}}{Significance Level}

\item{\code{min.obs}}{Minimum value to estimate the model (degree+1) x Groups + 1. (Default = 6).}

\item{\code{g}}{Integer. Number of genes taken in the regression fit.}

\item{\code{MT.adjust}}{Pvalue correction}

\item{\code{epsilon}}{convergence tolerance}

\item{\code{step.method}}{A character specifying the imputed step method for stepwise regression.}
}}

\keyword{classes}
