% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addParamClass.R
\docType{class}
\name{addParamClass}
\alias{addParamClass}
\alias{addParamClass-class}
\title{Class "addParamClass"}
\description{
Class "addParamClass"
}
\section{Slots}{

\describe{
\item{\code{bin_pseudotime_colname}}{A character representing the name of the column for binned pseudotime values.}

\item{\code{path_prefix}}{A character representing the prefix for path labeling.}

\item{\code{root_label}}{A character representing the label for the root of the tree.}

\item{\code{pseudotime_colname}}{A character representing the name of the column for pseudotime values.}

\item{\code{binning}}{A character representing the binning method.}

\item{\code{bin_method}}{A character representing the algorithm used for binning.}

\item{\code{path_colname}}{A character representing the name of the column for path values.}

\item{\code{bin_colname}}{A character representing the name of the column for bin values.}

\item{\code{bin_size_colname}}{A character representing the name of the column for bin sizes.}

\item{\code{bin_members_colname}}{A character representing the name of the column for bin members.}
}}

\keyword{classes}
