% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_monocle3_cds.R
\name{annotate_monocle3_cds}
\alias{annotate_monocle3_cds}
\title{Annotate Monocle3 Object}
\usage{
annotate_monocle3_cds(
  cds,
  reduction_method = "umap",
  path_prefix = "Path",
  root_label = "root",
  pseudotime_colname = "Pseudotime",
  path_colname = path_prefix,
  verbose = TRUE
)
}
\arguments{
\item{cds}{A Monocle3 CellDataSet (CDS).}

\item{reduction_method}{The dimensionality reduction method used. Default is
"umap". Monocle3 currently supports "UMAP" only for most procedure.}

\item{path_prefix}{Prefix used to annotate the paths. (Default is "Path").}

\item{root_label}{Label used to annotate root cells. (Default is "root").}

\item{pseudotime_colname}{Name of the column in `cell.metadata` generated using
\code{\link[SingleCellExperiment]{colData}} storing information for Pseudotime.
(Default is "Pseudotime")}

\item{path_colname}{Name of the column in `cell.metadata` generated using
\code{\link[SingleCellExperiment]{colData}} storing information for Path. 
(Default is `path_prefix`)}

\item{verbose}{Print detailed output in the console. (Default is TRUE)}
}
\value{
Annotated CDS object with end-point/path information:
}
\description{
`annotate_monocle3_cds()` annotates the Monocle3's CellDataSet object. It uses
the principal points, root cells, minimum spanning tree, to update the CDS
with end-point information that is used as path information.
}
\details{
The function first extracts the principal points and root cells from
the CDS. It then calculates the minimum spanning tree and the network matrix,
which contains the cells along each path from root to endpoint. Finally, it
updates the CDS with the endpoint data and the updated cell metadata.
}
\examples{
\dontrun{
# Assuming 'cds' is a Monocle3 CellDataSet (CDS)
result <- annotate_monocle3_cds(cds,
  reduction_method = "umap",
  path_prefix = "Path",
  root_label = "root",
  path_colname = path_prefix,
  verbose = TRUE
)
}

}
\seealso{
\code{\link[monocle3]{principal_graph_aux}}, \code{\link[monocle3]{principal_graph}},
\code{\link[igraph]{shortest_paths}}, \code{\link[monocle3]{pseudotime}}
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
