% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_sce.R
\name{annotate_sce}
\alias{annotate_sce}
\title{Annotate `SingleCellExperiment` class object with pseudotime and path
information.}
\usage{
annotate_sce(
  sce,
  ptime_col = "Pseudotime",
  path_prefix = "Path",
  root_label = "root",
  path_col = path_prefix,
  exist_ptime_col = NULL,
  exist_path_col = NULL,
  labels_exist = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{sce}{A `SingleCellExperiment` object to be annotated.}

\item{ptime_col}{A character string representing the column name
for inferred Pseudotime values. (Default is "Pseudotime")}

\item{path_prefix}{Prefix used to annotate the paths. (Default is "Path").}

\item{root_label}{Label used to annotate root cells. (Default is "root").}

\item{path_col}{A character string representing the column name for branching
path assignment. (Default is `path_prefix`)}

\item{exist_ptime_col}{The name of an existing pseudotime column
to be replaced (if not NULL).}

\item{exist_path_col}{The name of an existing path column to be replaced
(if not NULL).}

\item{labels_exist}{Logical indicating whether if the existing column names
should be overwritten. (Default is TRUE)}

\item{verbose}{Print detailed output in the console. (Default is TRUE)}
}
\value{
A \code{\link{SingleCellExperiment}} object with updated cell metadata.
}
\description{
`annotate_sce()` annotates a `SingleCellExperiment` class object with pseudotime
and path information in its `cell.metadata` generated using `colData` from the
\pkg{SingleCellExperiment} package.
}
\seealso{
\pkg{SingleCellExperiment} package.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
\keyword{internal}
