% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_scmp.R
\name{as_scmp}
\alias{as_scmp}
\title{Convert 'Cell Dataset' or 'SingleCellExperiment' object to scmpClass}
\usage{
as_scmp(
  object,
  from = "cds",
  path_prefix = "Path",
  root_label = "root",
  pseudotime_colname = "Pseudotime",
  path_colname = path_prefix,
  verbose = TRUE,
  additional_params = list(overwrite_labels = TRUE)
)
}
\arguments{
\item{object}{An S4 object of class `cell_data_set` or `SingleCellExperiment`.}

\item{from}{Character string specifying the class of 'object'. Use "cds" for
`cell_data_set` class and "sce" for `SingleCellExperiment` class.}

\item{path_prefix}{Prefix used to annotate the paths. (Default is "Path").}

\item{root_label}{Label used to annotate root cells. (Default is "root").}

\item{pseudotime_colname}{Name of the column in `cell.metadata` generated using
`colData` from the \pkg{SingleCellExperiment} package, storing information for Pseudotime.
(Default is "Pseudotime")}

\item{path_colname}{Name of the column in `cell.metadata` generated using
`colData` from the \pkg{SingleCellExperiment} package, storing information for Path.
(Default is `path_prefix`)}

\item{verbose}{Print detailed output in the console. (Default is TRUE)}

\item{additional_params}{A named list of additional parameters. See details.}
}
\value{
An instance of the 'scmpClass'.
}
\description{
`as_scmp()` converts a cell_data_set object from Monocle or a SingleCellExperiment object from Slingshot to an instance of the scmpClass object.
}
\details{
Additional Details
}
\examples{
\dontrun{
scmpObj <- as_scmp(object,
  from = "cds", path_prefix = "Path",
  root_label = "root",
  path_colname = path_prefix,
  verbose = TRUE
)
}

}
\seealso{
`colData` from the \pkg{SingleCellExperiment} package, `new_cell_data_set` function in \pkg{monocle3}
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
