% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_scmp_object_conversion.R
\name{as_scmp}
\alias{as_scmp}
\title{Convert cell_data_set or SingleCellExperiment to scMaSigProClass}
\usage{
as_scmp(object, from = "cell_data_set", verbose = T)
}
\arguments{
\item{object}{An S4 object of class 'cell_data_set' or 'SingleCellExperiment'.}

\item{from}{Character string specifying the class of 'object'. Can be either "cell_data_set" or "sce".}
}
\value{
An instance of the 'scMaSigProClass'.
}
\description{
This function converts a cell_data_set object from Monocle or a SingleCellExperiment object to an instance of
the scMaSigProClass. Currently, only Monocle's cell_data_set and SingleCellExperiment are supported.
}
\examples{
\dontrun{
scmpObj <- as_scmp(object, from = "sce")
}

}
