% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{bAlloc,ScMaSigPro-method}
\alias{bAlloc,ScMaSigPro-method}
\title{Set or get the Assignment Matrix.}
\usage{
\S4method{bAlloc}{ScMaSigPro}(object, value = "missing")
}
\arguments{
\item{object}{An object of class \code{\link{ScMaSigPro}}.}

\item{value}{The value of the branch Assignment Matrix.}
}
\value{
Modified `ScMaSigPro` object with new Assignment Matrix.
}
\description{
Set or get the Assignment Matrix.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
