% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{bAlloc}
\alias{bAlloc}
\title{Get or set the Branch Assignment Matrix}
\usage{
bAlloc(object, value = "missing")
}
\arguments{
\item{object}{An object of class \code{\link{ScMaSigPro}}.}

\item{value}{The value of the branch Assignment Matrix.}
}
\value{
Get the Assignment Matrix.
}
\description{
Get or set the Branch Assignment Matrix
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
