% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bin_size.R
\name{calc_bin_size}
\alias{calc_bin_size}
\title{Calculate Bin Size Function}
\usage{
calc_bin_size(x)
}
\arguments{
\item{x}{A data frame containing the "cluster.members" column.}
}
\value{
A numeric value representing the size of the bin (number of elements 
in the "cluster.members" column).
}
\description{
This function calculates the size of a bin based on the number of elements in 
the "cluster.members" column of the input data frame.
}
\examples{
# Sample data frame
data <- data.frame(
  cluster.members = c("A|B|C", "D|E", "F|G|H|I")
)

# Calculate the bin size using the calc_bin_size function
result <- calc_bin_size(data)
print(result)

}
\keyword{internal}
