% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{cDense<-,ScMaSigPro-method}
\alias{cDense<-,ScMaSigPro-method}
\title{Replacement method for cSparse}
\usage{
\S4method{cDense}{ScMaSigPro}(object) <- value
}
\arguments{
\item{object}{An object of class `ScMaSigPro`.}

\item{value}{The new value for the `colData` slot.}
}
\value{
Modified `ScMaSigPro` object.
}
\description{
Replacement method for cSparse
}
