% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{cDense}
\alias{cDense}
\title{Get or set the Sparse column data of a ScMaSigPro object}
\usage{
cDense(object, value = "missing")
}
\arguments{
\item{object}{An object of class `ScMaSigPro`.}

\item{value}{The new value for the `colData` slot to set. Optional for getting.}
}
\value{
`colData` when getting, modified `ScMaSigPro` object when setting.
}
\description{
Get or set the Sparse column data of a ScMaSigPro object
}
