% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_helpers.R
\name{create_range}
\alias{create_range}
\title{Create Range Function}
\usage{
create_range(
  x,
  bin_size_colname = "scmp_bin_size",
  bin_colname = "scmp_bin",
  verbose = TRUE
)
}
\arguments{
\item{x}{A data frame that should contain the following columns:
\describe{
  \item{bin}{A factor column representing the bin intervals in the format "[x, y]".}
  \item{bin_size}{A numeric column representing the bin size.}
  \item{binned_time}{A numeric column representing the binned time.}
}}
}
\value{
A numeric vector containing four elements:
  \describe{
    \item{lower_bound}{The lower bound of the bin interval.}
    \item{upper_bound}{The upper bound of the bin interval.}
    \item{bin_size}{The bin size.}
    \item{binned_time}{The binned time.}
  }
}
\description{
`create_range()` converts a factor column "bin" into a character vector, extracts
numeric range values from the character vector, and combines them with
additional columns "bin_size" and "binned_time" to return a numeric vector
representing the range information.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
\keyword{internal}
