% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{cSparse<-}
\alias{cSparse<-}
\title{Replacement method for cSparse}
\usage{
cSparse(object) <- value
}
\arguments{
\item{object}{An object of class `scmp`.}

\item{value}{The new value to set in the `colData` slot.}
}
\value{
Returns the modified `scmp` object.
}
\description{
This method allows replacing the `colData` of the `sparse` slot
of a `scmp` object.
}
