% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{cSparse}
\alias{cSparse}
\title{Get or set the sparse column data of a scmp object}
\usage{
cSparse(object, value = "missing")
}
\arguments{
\item{object}{An object of class `scmp`.}

\item{value}{The new value for the `colData` slot to set. Optional for getting.}
}
\value{
`colData` when getting, modified `scmp` object when setting.
}
\description{
Get or set the sparse column data of a scmp object
}
