% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{cSparse}
\alias{cSparse}
\title{Get the data for Sparse Slot.}
\usage{
cSparse(object, value = "missing")
}
\arguments{
\item{object}{An object of class \code{\link{ScMaSigPro}}.}

\item{value}{The new value for the `colData` slot of Sparse. (When Setting)}
}
\value{
`colData` for the `Sparse` Slot.
}
\description{
Get or set cell level metadata for the Sparse slot of an
\code{\link{ScMaSigPro}} object.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
