% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eDense,ScMaSigPro-method}
\alias{eDense,ScMaSigPro-method}
\title{Retrieve Expression Data from Dense Slot}
\usage{
\S4method{eDense}{ScMaSigPro}(object, i = "bulk.counts", ...)
}
\arguments{
\item{object}{An \code{\link{ScMaSigPro}} object from which to retrieve
the expression data.}

\item{i}{The name of the dataset to retrieve. The default value is "bulk.counts".}

\item{...}{Additional arguments (not used currently).}
}
\value{
Returns an expression matrix for the specified dataset from the Dense slot.
}
\description{
Retrieves expression data stored in the Dense slot of an 
\code{\link{ScMaSigPro}} object.
}
\seealso{
\code{\link{eDense<-}} for setting expression data in the Dense slot.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
