% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eDense<-,ScMaSigPro-method}
\alias{eDense<-,ScMaSigPro-method}
\title{Set eDense value for ScMaSigPro objects (Not Implemented)}
\usage{
\S4method{eDense}{ScMaSigPro}(object, i) <- value
}
\arguments{
\item{object}{The ScMaSigPro object.}

\item{i}{object name}

\item{value}{The value to set.}
}
\value{
None
}
\description{
Method to set the eDense value for ScMaSigPro class objects.
This method is currently not implemented.
}
