% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eDense<-}
\alias{eDense<-}
\title{Set Expression Counts in Dense Slot}
\usage{
eDense(object, i = "bulk.counts") <- value
}
\arguments{
\item{object}{An object of class \code{\link{ScMaSigPro}}.}

\item{i}{The name of the dataset for which to set expression counts.
Default is "bulk.counts".}

\item{value}{The new expression matrix to be set for the specified dataset.}
}
\value{
The modified \code{\link{ScMaSigPro}} object with updated expression
data in the Dense slot.
}
\description{
`eDense<-` is a generic function for updating or setting expression counts in the Dense slot
of an \code{\link{ScMaSigPro}} object.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
