% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edesignClass.R
\docType{class}
\name{edesignClass}
\alias{edesignClass}
\alias{edesignClass-class}
\title{edesignClass}
\description{
An S4 class to represent an edesign object with associated data.
This class contains three slots: dis, groups.vector, and edesign.
}
\section{Slots}{

\describe{
\item{\code{dis}}{A data frame.}

\item{\code{groups.vector}}{A character vector.}

\item{\code{edesign}}{A data frame.}

\item{\code{poly_degree}}{Polynomial degree}
}}

\section{Validity}{

  Valid objects must have:
  \itemize{
    \item{dis}{A valid data frame.}
    \item{groups.vector}{A valid character vector.}
    \item{edesign}{A valid data frame.}
  }
}

\keyword{classes}
