% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eSparse,ScMaSigPro-method}
\alias{eSparse,ScMaSigPro-method}
\title{Retrieve Expression Counts from Sparse Slot}
\usage{
\S4method{eSparse}{ScMaSigPro}(object, i = "counts", ...)
}
\arguments{
\item{object}{An \code{\link{ScMaSigPro}} object from which to retrieve expression data.}

\item{i}{The name of the assay to retrieve expression counts for.
The default value "counts" is a placeholder that results in the method returning
the first assay's data found in the Sparse slot.}

\item{...}{Additional arguments (not used currently).}
}
\value{
An expression matrix for the specified (or first, by default) assay 
from the Sparse slot.
}
\description{
Retrieves expression counts from the Sparse slot of an 
\code{\link{ScMaSigPro}} object.
}
\seealso{
\code{\link{eSparse<-}} for setting expression counts in the Sparse slot.
}
