% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eSparse<-,ScMaSigPro-method}
\alias{eSparse<-,ScMaSigPro-method}
\title{Set eSparse value for ScMaSigPro objects (Not Implemented)}
\usage{
\S4method{eSparse}{ScMaSigPro}(object, i) <- value
}
\arguments{
\item{object}{The ScMaSigPro object.}

\item{i}{object name}

\item{value}{The value to set.}
}
\value{
None
}
\description{
Method to set the eSparse value for ScMaSigPro class objects.
This method is currently not implemented.
}
