% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eSparse<-}
\alias{eSparse<-}
\title{Set eSparse value for an object}
\usage{
eSparse(object, i) <- value
}
\arguments{
\item{object}{The object to be modified.}

\item{value}{The value to be set for eSparse.}
}
\value{
None
}
\description{
`eSparse<-` is a generic function for setting eSparse value in an object.
Currently, this functionality is not implemented for all object types.
}
