% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpGenerics.R
\name{eSparse}
\alias{eSparse}
\title{Get Value for expression counts.}
\usage{
eSparse(object, value = "missing")
}
\arguments{
\item{object}{An object of class \code{\link{ScMaSigPro}}.}

\item{value}{The value to be set. (If Setting)}
}
\value{
The expression matrix for Sparse slot.
}
\description{
`eSparse<-` is a generic function for getting expression counts for Sparse
in an \code{\link{ScMaSigPro}}.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
