% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScMaSigProClass.R
\docType{class}
\name{Estimates}
\alias{Estimates}
\alias{Estimates-class}
\title{Estimates}
\description{
S4 class to store results of the polynomial term selection from full model.
}
\section{Slots}{

\describe{
\item{\code{significance_matrix}}{A matrix storing p-values for each gene per
polynomial term, R2 and global p-values.}

\item{\code{coefficient_matrix}}{A matrix storing beta estimates for each polynomial
term.}

\item{\code{path_coefficient_matrix}}{A matrix with the coefficients of each
branching path.}

\item{\code{t_score_matrix}}{A matrix storing the t-scores for each polynomial term.}

\item{\code{path}}{A character vector containing the branching path.}

\item{\code{influential}}{A matrix with genes containing influential data.}
}}

\keyword{classes}
