% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.intervals.R
\name{extract.intervals}
\alias{extract.intervals}
\title{Extract Intervals}
\usage{
extract.intervals(time.vector, nBins = 1, bin, bin.size, lbound, ubound)
}
\arguments{
\item{time.vector}{}

\item{nBins}{}

\item{bin}{}

\item{bin.size}{}

\item{ubond}{}
}
\description{
Extract Intervals
}
\keyword{internal}
