% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_helpers.R
\name{extract_fitting}
\alias{extract_fitting}
\title{Extract Fitting}
\usage{
extract_fitting(
  reg,
  lmf,
  model.glm.0,
  dis,
  family,
  name,
  vars.in,
  alfa,
  influ.info
)
}
\arguments{
\item{reg}{Regression Result form `stats::glm()`}

\item{lmf}{Full model object generated from the `stats::glm()`}

\item{model.glm.0}{Intercept model object generated from the `stats::glm()`}

\item{dis}{Dataframe for the predictors}

\item{family}{Family used for modelling the data.}

\item{name}{description}

\item{vars.in}{description}

\item{alfa}{Significance threshold (Q-Value)}

\item{influ.info}{description}
}
\description{
Extract Fitting
}
\keyword{internal}
