% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_helpers.R
\name{extract_interval}
\alias{extract_interval}
\title{Extract Intervals}
\usage{
extract_interval(time.vector, nBins = 1, bin, bin.size, lbound, ubound)
}
\arguments{
\item{time.vector}{Numeric vector of Pseudotime values.}

\item{nBins}{Expected number of bins.}

\item{bin}{Column name for the bin column.}

\item{bin.size}{Column name for the bin size column.}

\item{lbound}{Column name for the lower bound column.}

\item{ubound}{Column name for the upper bound column.}
}
\description{
Extract Intervals
}
\keyword{internal}
