% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_inspection.R
\name{glm_inspection}
\alias{glm_inspection}
\title{Visually Inspect Model Fit}
\usage{
glm_inspection(
  glm.model,
  plot.type = "fitted_vs_pearson",
  point.color = "blue",
  point.shape = 3,
  bar_width = 0.05,
  point.alpha = 0.8,
  ref.line.y.intercept = 0,
  func = "glm",
  input.data = NULL
)
}
\arguments{
\item{glm.model}{A model from the output of \link[stats]{glm}.}

\item{plot.type}{Plot type. Currently supported options are:
- "fitted_vs_pearson": Fitted values vs Pearson residuals plot.
- "cooks_vs_obs": Cook's distance vs observations bar plot.
- "pearson_vs_obs": Pearson residuals vs observed values plot.
- "fitted_vs_obs": Fitted values vs observed values plot.
- "std_vs_fitted": Standard residuals vs fitted values plot.}

\item{point.color}{Point colour for the plots.}

\item{point.shape}{Point shape for the plots.}

\item{bar_width}{Bar width for the Cook's distance plot.}

\item{point.alpha}{Point alpha for the plots.}

\item{ref.line.y.intercept}{Y intercept of the reference line in some plots.}

\item{func}{Which function is used (options: "glm" or "zeroinfl").}

\item{input.data}{No information.}
}
\description{
This function is used to visually inspect the model fit of a generalized linear
model (GLM) and provides different types of plots for validation purposes.
}
\keyword{internal}
