% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_select_path.R
\name{m3_select_path}
\alias{m3_select_path}
\title{Subset a CDS object interactively with Shiny}
\usage{
m3_select_path(
  cdsObj,
  redDim = "umap",
  annotation_col = "cell.type",
  pseudotime_col = "Pseudotime",
  path_col = "Path",
  use_shiny = TRUE,
  plot_purity = FALSE,
  m3_pp = list(root_pp = c(""), path1_pp = c(""), path2_pp = c(""), path1_name = NULL,
    path2_name = NULL)
)
}
\arguments{
\item{cdsObj}{An cdsObject of class `scmp`. This cdsObject will be checked
to ensure it's the right type.}

\item{redDim}{Dimension to use for the plot}

\item{annotation_col}{A character vector indicating the paths to be selected.}

\item{pseudotime_col}{Name of the column with Pseudotime}

\item{path_col}{Name of the column with Path}

\item{use_shiny}{description}

\item{plot_purity}{description}

\item{m3_pp}{description}
}
\value{
A `scmp` object, subsetted based on the specified paths.
}
\description{
Subset a CDS object interactively with Shiny
}
