% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_select_path.R
\name{m3_select_path}
\alias{m3_select_path}
\title{Select branching paths from a 'Cell Dataset' object from Monocle3}
\usage{
m3_select_path(
  cdsObj,
  redDim = "umap",
  annotation_col = "cell_type",
  pseudotime_col = "Pseudotime",
  path_col = "Path",
  use_shiny = TRUE,
  plot_purity = FALSE,
  m3_pp = list(root_pp = c(""), path1_pp = c(""), path2_pp = c(""), path1_name = NULL,
    path2_name = NULL)
)
}
\arguments{
\item{cdsObj}{An cdsObject from Monocle3.}

\item{redDim}{Dimension to use for the plot. (Default is "umap")}

\item{annotation_col}{Name of the column in `cell.metadata` storing
annotations. (Default is "cell_type")}

\item{pseudotime_col}{The name used for the Pesudotime column in the scmpObject.
(Default is "Pseudotime")}

\item{path_col}{The name used for the branching path column in the  scmpObject.
(Default is "Path")}

\item{use_shiny}{Enable the selection for shiny-selection wizard.
(Default is TRUE)}

\item{plot_purity}{Plot a bar plot, showing the the principal points against
the of `annotation_col`. This will show the count of annotations for each of
the principal points. (Default is TRUE)}

\item{m3_pp}{A list containing the character vectors for principal points ("Y_")
to be used for the branch selection.}
}
\value{
A `scmp` object, subsetted based on the specified paths.
}
\description{
`m3_select_path()` helps select branching paths from a from a 'Cell Dataset'
object from Monocle3 TI Analysis. This function also has an in-built shiny app
thats enabled interactive selection.
}
