% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m3_select_path.R
\name{m3_select_path}
\alias{m3_select_path}
\title{Select branching paths from a 'Cell Dataset' object from Monocle3}
\usage{
m3_select_path(
  cds,
  latent_space = "umap",
  anno_col = "cell_type",
  ptime_col = "Pseudotime",
  path_col = "Path",
  use_shiny = TRUE,
  plot_purity = FALSE,
  m3_pp = list(root_pp = c(""), path1_pp = c(""), path2_pp = c(""), path1_name = NULL,
    path2_name = NULL)
)
}
\arguments{
\item{cds}{An object from Monocle3.}

\item{latent_space}{Latent dimension to use for the plotting.
(Default is "umap")}

\item{anno_col}{A character string representing the column name for cell level
metadata containing cell level annotations. (Default is "cell_type").}

\item{ptime_col}{A character string representing the column name
for inferred Pseudotime values. (Default is "Pseudotime")}

\item{path_col}{A character string representing the column name for branching
path assignment. (Default is `path_prefix`)}

\item{use_shiny}{Enable the selection for shiny-selection wizard.
(Default is TRUE)}

\item{plot_purity}{Plot a bar plot, showing the the principal points against
the of `anno_col`. This will show the count of annotations for each of
the principal points. (Default is TRUE)}

\item{m3_pp}{A list containing the character vectors for principal
points ("Y_") to be used for the branch selection. See examples.}
}
\value{
An object of class \code{\link{ScMaSigPro}},
subsetted based on the specified paths.
}
\description{
`m3_select_path()` helps select branching paths from a from a 'Cell Dataset'
object from Monocle3 TI Analysis. This function also has an in-built shiny app
thats enabled interactive selection.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
