% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.bulk.count.R
\name{make.pseudobulk.counts}
\alias{make.pseudobulk.counts}
\title{Create Pseduo-bulk Counts}
\usage{
make.pseudobulk.counts(
  scmpObject,
  bin_members_colname = "scmp_bin_members",
  bin_colname = "scmp_bin",
  assay_name = "counts",
  pseudo_bulk_profile,
  cluster_count_by = "sum"
)
}
\arguments{
\item{scmpObject}{object of Class scMaSigPro. See \code{\link{scMaSigProClass}} 
for more details.}

\item{bin_members_colname}{Name of the column in the 'compressed_cell_metadata'
storing information about the members of the bins. (Default is 'scmp_bin_members').}

\item{bin_colname}{Name of the column in the 'compressed_cell_metadata'
storing information about the bin labels. (Default is 'scmp_bin').}

\item{assay_name}{Name of the Assay in the assay_name object from which retrieve the counts.
(Default = "counts").}

\item{pseudo_bulk_profile}{A data.frame generated using \code{\link{make.pseudobulk.design}}.}

\item{cluster_count_by}{A character string specifying the method to use to 
aggregate counts within each cluster. Available options are 'mean' or 'sum'. (Default = "sum").}
}
\value{
A data.frame. The data frame includes pseudo bulk counts with each row being a gene and each column being a bin.
}
\description{
`make.bulk.counts()` creates a dataframe of pseudo bulk counts from single cell counts. It does this by either taking the mean or sum of counts across clusters in each bin, depending on the specified method.
}
\details{
The function operates by iterating over each row of the pseudo_bulk_profile. 
For each bin, it identifies the cells that belong to the bin and selects their 
counts from the counts data frame. It then calculates the mean or sum of these 
counts (depending on the specified method), and adds these to a new data frame of pseudo bulk counts. 
The result is a pseudo bulk counts data frame where each row is a gene and each column is a bin.
}
\examples{
\dontrun{
make.pseudobulk.counts(
  counts = sc_counts,
  pseudo_bulk_profile = pb_profile,
  cluster_count_by = "mean"
)
}

}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
