% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.pseudobulk.design.R
\name{make.pseudobulk.design}
\alias{make.pseudobulk.design}
\title{Create Pseduo-bulk metadata}
\usage{
make.pseudobulk.design(
  scmpObject,
  path_colname = scmpObject@addParams@path_colname,
  bin_colname = "scmp_bin",
  bin_size_colname = "scmp_bin_size",
  bin_members_colname = "scmp_bin_members",
  bin_pseudotime_colname = scmpObject@addParams@bin_pseudotime_colname
)
}
\arguments{
\item{scmpObject}{object of Class scMaSigPro. See \code{\link{scMaSigProClass}}
for more details.}

\item{path_colname}{Name of the column in `cell.metadata` storing information
for Path. Generated using \code{\link[SingleCellExperiment]{colData}}. (Default
is `path_prefix`).}

\item{bin_colname}{Name of the column in the 'compressed_cell_metadata',
storing information about the bin labels. (Default is 'scmp_bin').}

\item{bin_size_colname}{Name of the column in the 'compressed_cell_metadata'
storing information about the size of the bins. (Default is 'scmp_bin_size').}

\item{bin_members_colname}{Name of the column in the 'compressed_cell_metadata'
storing information about the members of the bins. (Default is 'scmp_bin_members').}

\item{bin_pseudotime_colname}{Name of the column in the 'compressed_cell_metadata'
storing information about the binned pseudotime. (Default is 'scmp_binned_pseudotime').}
}
\value{
A data.frame containing the pseudobulk profile. The data frame includes the following columns:
- 'binnedTime': The time bin.
- 'cluster.members': The cells grouped in the bin.
- 'bin': The bin identifier.
- 'path': The path identifier.
- 'bin.size': The size of the bin (number of cells in the bin).
- Columns for each path in paths.vector, with binary values indicating whether the row belongs to the path.
}
\description{
`make.bulk.design()` creates a pseudobulk profile data frame from the design file.
It groups cells into bins for each path from a given set of paths and
calculates the bin size.
}
\details{
This function operates by iterating over the specified paths. For each path, it:
- Filters the design file to only include cells that belong to the path.
- Groups the cells into bins based on the binned time column.
- Calculates the size of each bin.
- Appends the resulting data frame to the pseudobulk profile.

After processing all paths, the function adds binary columns for each path to the pseudobulk profile.
}
\examples{
\dontrun{
make.pseudobulk.design(compressed_cell_metadata = df, path_colname = "path1", binned_pseudotime_column = "binnedTime")
}

}
\seealso{
\code{\link{calc_bin_size}}
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
