% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScMaSigProClass.R
\docType{class}
\name{MatrixDesign}
\alias{MatrixDesign}
\alias{MatrixDesign-class}
\title{MatrixDesign}
\description{
S4 class to store Branching Path Assignments and Prediction Matrix
}
\section{Slots}{

\describe{
\item{\code{predictor}}{A data frame containing the design matrix for the experiment.}

\item{\code{groups.vector}}{A character vector specifying the experimental group to which
each variable belongs.}

\item{\code{alloc}}{A data frame describing the detailed experimental design. Rows must contain
cells and columns experiment descriptors. The matrix must be binarized.}
}}

\section{Validity}{

  Valid objects must have:
  \itemize{
    \item{predictor}{A valid data frame.}
    \item{groups.vector}{A valid character vector.}
    \item{alloc}{A valid data frame.}
  }
}

\keyword{classes}
