% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBinTile.R
\name{plotBinTile}
\alias{plotBinTile}
\title{Plot Bin Sizes Across Binned Time and Paths}
\usage{
plotBinTile(
  scmpObj,
  path_col = scmpObj@Parameters@path_col,
  bin_size_col = scmpObj@Parameters@bin_size_col,
  bin_ptime_col = scmpObj@Parameters@bin_ptime_col
)
}
\arguments{
\item{scmpObj}{An object of class \code{\link{ScMaSigPro}}.}

\item{path_col}{A character string representing the column name for branching
path assignment in 'Sparse' or 'Dense' slot.}

\item{bin_size_col}{A character string representing the name of the column in
which bin sizes per bin are stored. (Default is "scmp_bin_size").}

\item{bin_ptime_col}{A character string representing the column name
for binned Pseudotime values in 'Dense' data.
(Default is "scmp_binned_pseudotime").}
}
\value{
ggplot2 plot object.
}
\description{
This function generates plots to visualize the Dense slot cell metadata
from a ScMaSigPro object. It produces tile plot to display the bin sizes
across different binned time intervals and paths.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
