% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBinTile.R
\name{plotBinTile}
\alias{plotBinTile}
\title{plotTile: Plot Bin Sizes Across Binned Time and Paths}
\usage{
plotBinTile(
  scmpObj,
  path_colname = scmpObj@param@path_colname,
  bin_size_colname = scmpObj@param@bin_size_colname,
  bin_pseudotime_colname = scmpObj@param@bin_pseudotime_colname
)
}
\arguments{
\item{scmpObj}{A SingleCellExperiment object with an additional slot '@compress.sce' that contains compression information.}

\item{path_colname}{A logical flag indicating whether to add a tile (heatmap) plot alongside the bar plot. Default is TRUE.}

\item{bin_size_colname}{A title of the barplot}

\item{bin_pseudotime_colname}{description}
}
\value{
A bar plot and, optionally, a tile (heatmap) plot, visualizing the bin sizes across different binned time and paths.
If add_tile is TRUE, returns a combined ggplot object with both plots; otherwise, only the bar plot is printed.
}
\description{
This function generates plots to visualize the compressed data from a SingleCellExperiment object.
It produces a bar plot and, optionally, a tile (heatmap) plot to display the bin sizes across different
binned time intervals and paths.
}
