% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIntersect.R
\name{plotIntersect}
\alias{plotIntersect}
\title{Generate UpSet Plot}
\usage{
plotIntersect(
  scmpObj,
  min_intersection_size = 1,
  keep_empty_groups = FALSE,
  width_ratio = 0.1,
  show_sets_size = FALSE,
  return = FALSE
)
}
\arguments{
\item{scmpObj}{An object of class \code{\link{ScMaSigPro}}.}

\item{min_intersection_size}{Minimal number of observations in an intersection
for it to be included.}

\item{keep_empty_groups}{Whether empty sets should be kept (including sets
which are only empty after filtering by size)}

\item{width_ratio}{Ratio of the overall set size width to intersection matrix
width.}

\item{show_sets_size}{The overall set sizes plot, e.g. from upset_set_size()}

\item{return}{If set to true, it will return dataframe from the UpSetR::fromList().
(Default is TRUE)}
}
\value{
upset object for 'UpSetR'.
}
\description{
Generate UpSet Plot on Intersection of Significant Genes from scMaSigPro
object. It is a wrapper `UpSetR::upset`.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
