% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrend.R
\name{plotTrend}
\alias{plotTrend}
\title{Plot Groups Function}
\usage{
plotTrend(
  scmpObj,
  feature_id,
  xlab = "Pooled Pseudotime",
  ylab = "Pseudobulk Expression",
  smoothness = 0.01,
  logs = TRUE,
  logType = "log",
  pseudoCount = 1,
  significant = TRUE,
  summary_mode = "median"
)
}
\arguments{
\item{scmpObj}{object of class scmpObj}

\item{feature_id}{Name of the gene to be plotted. Should correspond to one of
the feature in the count table.}

\item{xlab}{X-axis label. Default is "Pooled Pseudotime".}

\item{ylab}{Y-axis label. Default is "Pseudobulk Expression".}

\item{smoothness}{How smooth the trend should be. Default is 0.01, setting to
higher values will result in more linear trends.}

\item{logs}{Whether to plot log of counts.}

\item{logType}{Log Available options 'log', 'log2', 'log10'}

\item{pseudoCount}{Add a pseudo-count before taking the log.}

\item{significant}{Default is FALSE. Set to TRUE to plot genes, that don't pass
R-Square threshold from 'sc.filter()'.}

\item{summary_mode}{description}
}
\value{
Generates a plot.
}
\description{
This function generates plots based on various parameters. It calculates the summary mode, colors, and other visual attributes to create a plot.
}
