% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrend.R
\name{plotTrend}
\alias{plotTrend}
\title{Plot trend of the single gene.}
\usage{
plotTrend(
  scmpObj,
  feature_id,
  xlab = "Pooled Pseudotime",
  ylab = "Pseudobulk Expression",
  smoothness = 0.01,
  logs = TRUE,
  logType = "log",
  pseudoCount = 1,
  significant = TRUE,
  summary_mode = "median",
  curves = TRUE,
  lines = FALSE,
  points = TRUE
)
}
\arguments{
\item{scmpObj}{An object of class \code{\link{ScMaSigPro}}.}

\item{feature_id}{Name of the gene to be plotted.}

\item{xlab}{X-axis label. (Default is "Pooled Pseudotime")}

\item{ylab}{Y-axis label. (Default is "Pseudobulk Expression")}

\item{smoothness}{How smooth the trend should be. Setting to
higher values will result in more linear trends. (Default is 0.01)}

\item{logs}{Whether to log transform counts. (Default is TRUE)}

\item{logType}{How to log transform the values. Available options 'log',
'log2', 'log10'. (Default is 'log')}

\item{pseudoCount}{Add a pseudo-count before taking the log. (Default is 1)}

\item{significant}{Plot gene only if the models are significant based on
\code{scMaSigPro::sc.filter()}. (Default is TRUE)}

\item{summary_mode}{Compress the expression values per replicate (if present)
per binned pseudotime point. Default is 'median'. Other option 'mean'}

\item{curves}{Whether to plot the fitted curves. (Default is TRUE)}

\item{lines}{Whether to plot the lines. (Default is FALSE)}

\item{points}{Whether to plot the points. (Default is TRUE)}
}
\value{
ggplot2 plot object.
}
\description{
Plot trend of the single gene across the binned pseudotime.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
