% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrendCluster.R
\name{plotTrendCluster}
\alias{plotTrendCluster}
\title{Plot Groups Function}
\usage{
plotTrendCluster(
  scmpObj,
  geneSet,
  xlab = "Pooled Pseudotime",
  ylab = "Pseudobulk Expression",
  cluster_by = "coeff",
  cluster_method = "hclust",
  hclust_agglo_method = "ward.D",
  kmeans_iter_max = 500,
  distance = "cor",
  summary_mode = "median",
  logs = TRUE,
  logType = "log",
  smoothness = 0.01,
  k = 9,
  includeInflu = TRUE,
  result = "plot",
  mclust_k = FALSE,
  verbose = FALSE,
  pseudoCount = 1,
  scale_counts = TRUE,
  significant = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{scmpObj}{object of class scmpObj}

\item{geneSet}{description}

\item{xlab}{X-axis label. Default is "Pooled Pseudotime".}

\item{ylab}{Y-axis label. Default is "Pseudobulk Expression".}

\item{cluster_by}{description}

\item{cluster_method}{description}

\item{hclust_agglo_method}{description}

\item{kmeans_iter_max}{description}

\item{distance}{description}

\item{summary_mode}{description}

\item{logs}{Whether to plot log of counts}

\item{logType}{Log type required}

\item{smoothness}{description}

\item{k}{description}

\item{includeInflu}{description}

\item{result}{description}

\item{mclust_k}{description}

\item{verbose}{description}

\item{pseudoCount}{description}

\item{scale_counts}{description}

\item{significant}{description}

\item{parallel}{description}
}
\value{
Generates a plot.
}
\description{
This function generates plots based on various parameters.
}
