% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrendCluster.R
\name{plotTrendCluster}
\alias{plotTrendCluster}
\title{Plot multiple trends of the multiple genes.}
\usage{
plotTrendCluster(
  scmpObj,
  xlab = "Pooled Pseudotime",
  ylab = "Pseudobulk Expression",
  plot = "counts",
  summary_mode = "median",
  logs = FALSE,
  logType = "log",
  smoothness = 1,
  includeInflu = TRUE,
  verbose = TRUE,
  pseudoCount = 1,
  significant = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{scmpObj}{object of class ScMaSigPro.}

\item{xlab}{X-axis label. (Default is "Pooled Pseudotime")}

\item{ylab}{Y-axis label. (Default is "Pseudobulk Expression")}

\item{plot}{Whether to plot 'coeff' or 'counts'. (Default is 'counts')}

\item{summary_mode}{Compress the expression values for per replicate 
(if present) per binned pseudotime point. Default is 'median'. Other option
'mean'}

\item{logs}{Whether to log transform counts. (Default is TRUE)}

\item{logType}{How to log transform the values. Available options 'log',
'log2', 'log10'. (Default is 'log')}

\item{smoothness}{How smooth the trend should be. Setting to
higher values will result in more linear trends. (Default is 0.01)}

\item{includeInflu}{Include gene only if it has influential data. 
(Default is TRUE)}

\item{verbose}{Print detailed output in the console. (Default is TRUE)}

\item{pseudoCount}{Add a pseudo-count before taking the log. (Default is 1)}

\item{significant}{Include gene only if the models are significant based on 
\code{\link{scMaSigPro::sc.filter()}}. (Default is TRUE)}

\item{parallel}{Use forking process to run parallelly. (Default is FALSE)
(Currently, Windows is not supported)}
}
\value{
ggplot2 plot object.
}
\description{
Plot trends of multiple genes (clustered) across the binned pseudotime.
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
