% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrendCluster.R
\name{plotTrendCluster}
\alias{plotTrendCluster}
\title{Plot Groups Function}
\usage{
plotTrendCluster(
  scmpObj,
  geneSet,
  xlab = "Pooled Pseudotime",
  ylab = "Pseudobulk Expression",
  cluster_by = "coeff",
  cluster_method = "hclust",
  logs = TRUE,
  logType = "log",
  smoothness = 0.01,
  k = 9,
  includeInflu = TRUE,
  distance = "cor",
  hclust.agglo_method = "ward.D",
  result = "plot",
  significant = FALSE
)
}
\arguments{
\item{scmpObj}{object of class scmpObj}

\item{geneSet}{description}

\item{xlab}{X-axis label. Default is "Pooled Pseudotime".}

\item{ylab}{Y-axis label. Default is "Pseudobulk Expression".}

\item{cluster_by}{description}

\item{cluster_method}{description}

\item{logs}{Whether to plot log of counts}

\item{logType}{Log type required}

\item{smoothness}{description}

\item{k}{description}

\item{includeInflu}{description}

\item{distance}{description}

\item{hclust.agglo_method}{description}

\item{result}{description}

\item{significant}{description}
}
\value{
Generates a plot.
}
\description{
This function generates plots based on various parameters.
}
