% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrendCluster.R
\name{plotTrendCluster}
\alias{plotTrendCluster}
\title{Plot Groups Function}
\usage{
plotTrendCluster(
  scmpObj,
  xlab = "Pooled Pseudotime",
  ylab = "Pseudobulk Expression",
  plot = "counts",
  summary_mode = "median",
  logs = FALSE,
  logType = "log",
  smoothness = 1,
  includeInflu = TRUE,
  verbose = TRUE,
  pseudoCount = 1,
  significant = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{scmpObj}{object of class scmpObj}

\item{xlab}{X-axis label. Default is "Pooled Pseudotime".}

\item{ylab}{Y-axis label. Default is "Pseudobulk Expression".}

\item{plot}{description}

\item{summary_mode}{description}

\item{logs}{Whether to plot log of counts}

\item{logType}{Log type required}

\item{smoothness}{description}

\item{includeInflu}{description}

\item{verbose}{description}

\item{pseudoCount}{description}

\item{significant}{description}

\item{parallel}{description}
}
\value{
Generates a plot.
}
\description{
This function generates plots based on various parameters.
}
