% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.cluster.trend.R
\name{sc.cluster.trend}
\alias{sc.cluster.trend}
\title{Cluster the counts or coefficients}
\usage{
sc.cluster.trend(
  scmpObj,
  geneSet,
  cluster_by = "coeff",
  cluster_method = "hclust",
  hclust.agglo.method = "ward.D",
  kmeans.iter.max = 500,
  hclust.distance = "cor",
  includeInflu = TRUE,
  mclust.k = FALSE,
  k = 9,
  verbose = FALSE,
  fill_na_by = "zero",
  fill_dim = "col"
)
}
\arguments{
\item{scmpObj}{object of class scmpObj}

\item{geneSet}{description}

\item{cluster_by}{description}

\item{cluster_method}{description}

\item{hclust.agglo.method}{description}

\item{kmeans.iter.max}{description}

\item{hclust.distance}{description}

\item{includeInflu}{description}

\item{mclust.k}{description}

\item{k}{description}

\item{verbose}{description}

\item{fill_na_by}{description}

\item{fill_dim}{description}
}
\value{
Generates a plot.
}
\description{
This function clusters the counts or coefficients
}
