% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.filter.R
\name{sc.filter}
\alias{sc.filter}
\title{Extract significant genes for sets of variables in time series gene expression experiments}
\usage{
sc.filter(
  scmpObj,
  rsq = 0.7,
  Q = scmpObj@addParams@Q,
  vars = c("all", "each", "groups"),
  significant.intercept = "dummy",
  term.Q = 0.05,
  includeInflu = TRUE
)
}
\arguments{
\item{scmpObj}{Object of Class \code{\link{scMaSigProClass}} in which the
\code{sc.T.fit} has been run.}

\item{rsq}{Cut-off level at the R-squared value for the stepwise regression fit.}

\item{Q}{overall model significance}

\item{vars}{Variables for which to extract significant genes. There are 3 possible values:
\itemize{
  \item \code{"all"}: generates one single matrix or gene list with all significant genes.
  \item \code{"each"}: generates as many significant genes extractions as variables in the general regression model.
  \item \code{"groups"}: generates a significant genes extraction for each experimental group.
}}

\item{significant.intercept}{Experimental groups for which significant intercept coefficients are considered.}

\item{term.Q}{Term wise significance
Only genes with R-squared more than 'rsq' are selected. (Default = 0.7).}

\item{includeInflu}{description}
}
\value{
A list with two elements:
\itemize{
  \item \code{summary}: A vector or matrix listing significant genes for the variables given by the function parameters.
  \item \code{sig.genes}: A list with detailed information on the significant genes found for the variables given by the function parameters.
  Each element of the list is also a list containing:
    \describe{
      \item{\code{sig.profiles}:}{Expression values of significant genes.}
      \item{\code{coefficients}:}{Regression coefficients of the adjusted models.}
      \item{\code{group.coeffs}:}{Regression coefficients of the implicit models of each experimental group.}
      \item{\code{sig.pvalues}:}{P-values of the regression coefficients for significant genes.}
      \item{\code{g}:}{Number of genes.}
      \item{\code{...}:}{Arguments passed by previous functions.}
    }
  }
}
\description{
`sc.filter()` creates lists of significant genes for a set of variables
whose significance value has been computed with the \code{sc.T.fit} function.
}
\details{
Refer to the function description for details on the arguments and their usage.
}
\examples{
# Example usage of the function can be placed here.

}
\references{
Conesa, A., Nueda M.J., Alberto Ferrer, A., Talon, T. (2006).
maSigPro: a Method to Identify Significant Differential Expression Profiles in Time-Course Microarray Experiments.
Bioinformatics, 22(9), 1096-1102. \url{https://doi.org/10.1093/bioinformatics/btl056}
}
\author{
Ana Conesa and Maria Jose Nueda (mj.nueda@ua.es)
}
\keyword{manip}
