% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.p.vector.R
\name{sc.p.vector}
\alias{sc.p.vector}
\title{Perform fitting with full model.}
\usage{
sc.p.vector(
  scmpObj,
  p_value = 0.05,
  mt_correction = "BH",
  min_na = 6,
  family = negative.binomial(theta = 10),
  epsilon = 1e-08,
  verbose = TRUE,
  offset = TRUE,
  parallel = FALSE,
  log_offset = FALSE,
  max_it = 100,
  link = "log"
)
}
\arguments{
\item{scmpObj}{An object of class \code{\link{ScMaSigPro}}.}

\item{p_value}{Significance level used for variable selection in the stepwise
regression.}

\item{mt_correction}{A character string specifying the p-value correction
method.}

\item{min_na}{Minimum values needed per gene across cells to estimate the
model.}

\item{family}{Distribution of the error term.}

\item{epsilon}{Model convergence tolerance.}

\item{verbose}{Print detailed output in the console. (Default is TRUE)}

\item{offset}{logical value specifying whether to use offset during fitting.}

\item{parallel}{Use forking process to run parallelly. (Default is FALSE)
(Currently, Windows is not supported)}

\item{log_offset}{A logical value specifying whether to take the logarithm of
the offsets.}

\item{max_it}{Maximum number of iterations to fit the model.}

\item{link}{Type of link function to use in the model. Default is "log".}

\item{Type}{of link function to use in the model. Default is "log".}
}
\value{
An object of class \code{\link{ScMaSigPro}}, with updated `Profile`
slot.
}
\description{
Performs a regression fit for each gene taking all variables present in the
model.
}
\references{
Conesa, A., Nueda M.J., Alberto Ferrer, A., Talon, T. 2006.
maSigPro: a Method to Identify Significant Differential Expression Profiles
in Time-Course Microarray Experiments. Bioinformatics 22, 1096-1102
}
\seealso{
\code{\link{VariableProfiles}}
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}, Ana Conesa and
Maria Jose Nueda, \email{mj.nueda@ua.es}
}
\keyword{models}
\keyword{regression}
