% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.p.vector.R
\name{sc.p.vector}
\alias{sc.p.vector}
\title{Make regression fit for time series gene expression experiments}
\usage{
sc.p.vector(
  scmpObj,
  Q = 0.05,
  MT.adjust = "BH",
  min.obs = 6,
  counts = FALSE,
  family = NULL,
  theta = 10,
  epsilon = 1e-05,
  item = "gene",
  verbose = TRUE,
  offset = T
)
}
\arguments{
\item{scmpObj}{matrix containing normalized gene expression data. Genes must be in rows and arrays in columns.}

\item{Q}{significance level. Default is 0.05.}

\item{MT.adjust}{argument to pass to \code{p.adjust} function indicating the method for multiple testing adjustment of p.value.}

\item{min.obs}{genes with less than this number of true numerical values will be excluded from the analysis.
Minimum value to estimate the model is (degree+1) x Groups + 1. Default is 6.}

\item{counts}{a logical indicating whether your data are counts.}

\item{family}{the distribution function to be used in the glm model.
It must be specified as a function: \code{gaussian()}, \code{poisson()}, \code{negative.binomial(theta)}...
If NULL, the family will be \code{negative.binomial(theta)} when \code{counts = TRUE} or \code{gaussian()} when \code{counts = FALSE}.}

\item{theta}{theta parameter for negative.binomial family.}

\item{epsilon}{argument to pass to \code{glm.control}, convergence tolerance in the iterative process to estimate the glm model.}

\item{item}{Name of the analyzed item to show on the screen while \code{sc.p.vector} is in process.}

\item{verbose}{Name of the analyzed item to show on the screen while \code{T.fit} is in process.}

\item{offset}{Whether ro use offset for normalization}
}
\value{
A list containing:
\item{SELEC}{matrix containing the expression values for significant genes}
\item{sc.p.vector}{vector containing the computed p-values}
\item{G}{total number of input genes}
\item{g}{number of genes taken in the regression fit}
\item{FDR}{p-value at FDR \code{Q} control when Benjamini & Hochberg (BH) correction is used}
\item{i}{number of significant genes}
\item{dis}{design matrix used in the regression fit}
\item{dat}{matrix of expression value data used in the regression fit}
\item{...}{additional values from input parameters}
}
\description{
\code{sc.p.vector} performs a regression fit for each gene taking all variables present in the model given by a regression matrix
and returns a list of FDR corrected significant genes.
}
\details{
\code{rownames(design)} and \code{colnames(data)} must be identical vectors
  and indicate array naming. \code{rownames(data)} should contain unique gene IDs.
  \code{colnames(design)} are the given names for the variables in the regression model.
}
\examples{
#### GENERATE TIME COURSE DATA
## generates n random gene expression profiles of a data set with
## one control plus 3 treatments, 3 time points and r replicates per time point.

# ... (add example usage of the function here)

}
\references{
Conesa, A., Nueda M.J., Alberto Ferrer, A., Talon, T. 2006.
maSigPro: a Method to Identify Significant Differential Expression Profiles in Time-Course Microarray Experiments.
Bioinformatics 22, 1096-1102
}
\seealso{
\code{\link{T.fit}}, \code{\link{lm}}
}
\author{
Ana Conesa and Maria Jose Nueda, \email{mj.nueda@ua.es}
}
\keyword{regression}
