% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.set.poly.R
\name{sc.set.poly}
\alias{sc.set.poly}
\title{Create predictors and set polynomial.  Adaption of maSigPro::make.design.matrix()}
\usage{
sc.set.poly(
  scmpObject,
  poly_degree = 2,
  bin_ptime_col = scmpObject@Parameters@bin_ptime_col,
  path_col = scmpObject@Parameters@path_col
)
}
\arguments{
\item{scmpObject}{A 'ScMaSigPro' object.}

\item{poly_degree}{Degree of the design matrix (default: 2).}

\item{bin_ptime_col}{Name of the time column.}

\item{path_col}{Name of the path column.}
}
\value{
Returns the 'scmpObject' with an updated 'design' slot.
}
\description{
Create predictors and set polynomial.  Adaption of maSigPro::make.design.matrix()
}
