% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.set.poly.R
\name{sc.set.poly}
\alias{sc.set.poly}
\title{Create design matrix for 'scMaSigProClass' object}
\usage{
sc.set.poly(
  scmpObject,
  poly_degree = 2,
  bin_pseudotime_colname = scmpObject@addParams@bin_pseudotime_colname,
  path_colname = scmpObject@addParams@path_colname
)
}
\arguments{
\item{scmpObject}{A 'scMaSigProClass' object.}

\item{poly_degree}{Degree of the design matrix (default: 2).}

\item{bin_pseudotime_colname}{Name of the time column.}

\item{path_colname}{Name of the path column.}
}
\value{
Returns the 'scmpObject' with an updated 'edesign' slot.
}
\description{
This function creates a design matrix using the 'compress.sce' slot of a 'scMaSigProClass' object.
It generates an 'edesignClass' object which is then stored in the 'edesign' slot of the 'scMaSigProClass' object.
}
