% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.T.fit.R
\name{sc.T.fit}
\alias{sc.T.fit}
\title{Makes a stepwise regression fit for time series gene expression experiments}
\usage{
sc.T.fit(
  scmpObj,
  step.method = "backward",
  Q = scmpObj@addParams@Q,
  nvar.correction = FALSE,
  family = scmpObj@distribution,
  epsilon = scmpObj@addParams@epsilon,
  offset = scmpObj@addParams@offset,
  verbose = TRUE,
  parallel = TRUE,
  useWeights = scmpObj@addParams@useWeights,
  useInverseWeights = scmpObj@addParams@useInverseWeights,
  logWeights = scmpObj@addParams@logWeights,
  logOffset = scmpObj@addParams@logOffset,
  max_it = scmpObj@addParams@max_it,
  globalTheta = scmpObj@addParams@globalTheta
)
}
\arguments{
\item{scmpObj}{Can either be a \code{\link{p.vector}} object or a matrix containing expression scmpObj with the same requirements as for
the \code{\link{p.vector}} function.}

\item{step.method}{Argument to be passed to the step function. Can be either \code{"backward"}, \code{"forward"}, \code{"two.ways.backward"}, or \code{"two.ways.forward"}.}

\item{Q}{Significance level used for variable selection in the stepwise regression.}

\item{nvar.correction}{Argument for correcting T.fit significance level. See details.}

\item{family}{The distribution function to be used in the glm model. It must be the same used in \code{p.vector}.}

\item{epsilon}{Argument to pass to \code{glm.control}, convergence tolerance in the iterative process to estimate the glm model.}

\item{offset}{Whether ro use offset for normalization}

\item{verbose}{Name of the analyzed item to show on the screen while \code{T.fit} is in process.}

\item{parallel}{description}

\item{useWeights}{Use bin size as weights}

\item{useInverseWeights}{description}

\item{logWeights}{description}

\item{logOffset}{description}

\item{max_it}{description}

\item{globalTheta}{description}
}
\value{
A list containing the following elements:
\item{sol}{Matrix for summary results of the stepwise regression. For each selected gene, the following values are given:
\itemize{
  \item p-value of the regression ANOVA
  \item R-squared of the model
  \item p-value of the regression coefficients of the selected variables
}}
\item{coefficients}{Matrix containing regression coefficients for the adjusted models.}
\item{group.coeffs}{Matrix containing the coefficients of the implicit models of each experimental group.}
\item{variables}{Variables in the complete regression model.}
\item{G}{Total number of input genes.}
\item{g}{Number of genes taken in the regression fit.}
\item{dat}{Input analysis scmpObj matrix.}
\item{dis}{Regression design matrix.}
\item{step.method}{Imputed step method for stepwise regression.}
\item{edesign}{Matrix of experimental design.}
\item{influ.info}{scmpObj frame of genes containing influential scmpObj.}
}
\description{
\code{T.fit} selects the best regression model for each gene using stepwise regression.
}
\details{
In the maSigPro approach, \code{\link{p.vector}} and \code{\link{T.fit}} are subsequent steps, meaning that significant genes are
first selected based on a general model, and then the significant variables for each gene are found by step-wise regression.

The step regression can be \code{"backward"} or \code{"forward"}, indicating whether the step procedure starts from the
model with all or none variables. With the \code{"two.ways.backward"} or \code{"two.ways.forward"} options, the variables are both allowed to get in and out.
At each step, the p-value of each variable is computed, and variables get in/out of the model when this p-value is
lower or higher than the given threshold \code{Q}. When \code{nvar.correction} is TRUE, the given significance level is corrected by the number of variables in the model.
}
\references{
{Conesa, A., Nueda M.J., Alberto Ferrer, A., Talon, T. 2006.
maSigPro: a Method to Identify Significant Differential Expression Profiles in Time-Course Microarray Experiments.
Bioinformatics 22, 1096-1102}
}
\seealso{
{\code{\link{p.vector}}, \code{\link{step}}}
}
\author{
{Ana Conesa and Maria Jose Nueda, \email{mj.nueda@ua.es}}
}
\keyword{models}
\keyword{regression}
