% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScMaSigProClass.R
\docType{class}
\name{ScMaSigPro}
\alias{ScMaSigPro}
\alias{ScMaSigPro-class}
\title{ScMaSigPro}
\description{
S4 class to store analysis of ScMaSigPro worflow. It stores results,
parameters and associated data. Inherits slots from \code{SingleCellExperiment}.
}
\section{Slots}{

\describe{
\item{\code{sparse}}{Object of Class SingleCellExperiment. See \pkg{SingleCellExperiment} for more details.}

\item{\code{profile}}{Object of Class VariableProfiles See \pkg{VariableProfiles} for more details.}

\item{\code{estimate}}{Object of Class Estimates See \code{\link{Estimates}} for more details.}

\item{\code{dense}}{Object of Class SingleCellExperiment. See \pkg{SingleCellExperiment} for more details.}

\item{\code{design}}{Object of Class MatrixDesign. See \code{\link{MatrixDesign}} for more details.}

\item{\code{param}}{Object of Class ParameterConfig. See \code{\link{ParameterConfig}} for more details.}

\item{\code{significant}}{ABC}
}}

\keyword{classes}
