% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpClass.R
\docType{class}
\name{scMaSigProClass}
\alias{scMaSigProClass}
\alias{scMaSigProClass-class}
\title{scMaSigProClass}
\description{
A class to represent the ScMaSigPro analysis results and associated data.
Inherits from \code{SingleCellExperiment}.
}
\section{Slots}{

\describe{
\item{\code{sce}}{Object of Class SingleCellExperiment. See \pkg{SingleCellExperiment} for more details.}

\item{\code{scPVector}}{Object of Class scPVectorClass See \pkg{scPVectorClass} for more details.}

\item{\code{scTFit}}{Object of Class scTFitClass. See \code{\link{scTFitClass}} for more details.}

\item{\code{compress.sce}}{ABC}

\item{\code{edesign}}{Object of Class edesignClass. See \code{\link{edesignClass}} for more details.}

\item{\code{siggenes}}{ABC}

\item{\code{addParams}}{Object of Class addParamClass. See \code{\link{addParamClass}} for more details.}
}}

\keyword{classes}
