% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpClass.R
\docType{class}
\name{scMaSigProClass}
\alias{scMaSigProClass}
\alias{scMaSigProClass-class}
\title{scMaSigProClass}
\description{
A class to represent the ScMaSigPro analysis results and associated data.
Inherits from \code{SingleCellExperiment}.
}
\section{Slots}{

\describe{
\item{\code{sparse}}{Object of Class SingleCellExperiment. See \pkg{SingleCellExperiment} for more details.}

\item{\code{scPVector}}{Object of Class scPVectorClass See \pkg{scPVectorClass} for more details.}

\item{\code{scTFit}}{Object of Class scTFitClass. See \code{\link{scTFitClass}} for more details.}

\item{\code{dense}}{ABC}

\item{\code{edesign}}{Object of Class edesignClass. See \code{\link{edesignClass}} for more details.}

\item{\code{param}}{Object of Class paramClass. See \code{\link{paramClass}} for more details.}

\item{\code{sig.genes}}{ABC}

\item{\code{distribution}}{The distribution function to be used in the glm model.}
}}

\keyword{classes}
