% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmpClass.R
\docType{class}
\name{scMaSigProClass}
\alias{scMaSigProClass}
\title{scMaSigProClass}
\description{
scMaSigProClass
}
\section{Slots}{

\describe{
\item{\code{covariate}}{Description of Slot.}

\item{\code{parameters}}{Description of Slot.}
}}

