% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.p.vector.class.R
\docType{class}
\name{scPVectorClass-class}
\alias{scPVectorClass-class}
\title{A class to store regression fit results for time series gene expression experiments}
\description{
The scPVectorClass is designed to hold the results of a regression fit for each gene in a time series gene expression experiment.
It contains slots to store various results, such as expression values for significant genes, computed p-values, adjusted p-values,
number of input genes, number of genes taken in the regression fit, and more. This class is useful for analyzing time course
microarray experiments and identifying significant differential expression profiles.
}
\section{Slots}{

\describe{
\item{\code{SELEC}}{Matrix containing the expression values for significant genes}

\item{\code{sc.p.vector}}{Matrix containing the computed p-values}

\item{\code{p.adjusted}}{Vector of FDR-adjusted p-values}

\item{\code{G}}{Total number of input genes}

\item{\code{g}}{Number of genes taken in the regression fit}

\item{\code{FDR}}{P-value at FDR \code{Q} control when Benjamini & Hochberg (BH) correction is used}

\item{\code{i}}{Number of significant genes}

\item{\code{dis}}{Design matrix used in the regression fit}

\item{\code{dat}}{Matrix of expression value data used in the regression fit}

\item{\code{min.obs}}{Minimum value to estimate the model (degree+1) x Groups + 1. Default is 6.}

\item{\code{Q}}{Significance level (default is 0.05)}

\item{\code{groups.vector}}{List containing groups information}

\item{\code{edesign}}{Experimental design data frame}

\item{\code{family}}{Distribution function to be used in the glm model. If NULL, the family will be \code{negative.binomial(theta)} when \code{counts = TRUE} or \code{gaussian()} when \code{counts = FALSE}.}
}}

\examples{
#### GENERATE TIME COURSE DATA
## generates n random gene expression profiles of a data set with
## one control plus 3 treatments, 3 time points and r replicates per time point.

# ... (add example usage of the function here)

}
\seealso{
\code{\link{T.fit}}, \code{\link{lm}}
}
\author{
Priyansh Srivastava <spriyansh29@gmail.com>
}
\keyword{regression}
