% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPath.m3.R
\name{selectPath.m3}
\alias{selectPath.m3}
\title{Subset a CDS object interactively with Shiny}
\usage{
selectPath.m3(
  cdsObj,
  redDim = "umap",
  annotation_col = "cell.type",
  pseudotime_col = "Pseudotime",
  path_col = "Path"
)
}
\arguments{
\item{cdsObj}{An cdsObject of class `scmp`. This cdsObject will be checked
to ensure it's the right type.}

\item{redDim}{Dimension to use for the plot}

\item{annotation_col}{A character vector indicating the paths to be selected.}

\item{pseudotime_col}{Name of the column with Pseudotime}

\item{path_col}{Name of the column with Path}
}
\value{
A `scmp` object, subsetted based on the specified paths.
}
\description{
Subset a CDS object interactively with Shiny
}
