% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinySelection.R
\name{shinySelect}
\alias{shinySelect}
\title{Function to interactively select nodes from shiny}
\usage{
shinySelect(
  trajectory_data,
  annotation_data,
  label_coords,
  pseudotime_colname,
  inputType = "Monocle3"
)
}
\arguments{
\item{trajectory_data}{Dataframe for trajectory}

\item{annotation_data}{Dataframe for cell UMAP}

\item{label_coords}{Datatframe for labels}

\item{inputType}{Input type}
}
\description{
Function to interactively select nodes from shiny
}
\keyword{internal}
