% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showFunctions.R
\name{showPoly}
\alias{showPoly}
\title{Show the terms of the polynomial term}
\usage{
showPoly(scmpObj)
}
\arguments{
\item{scmpObj}{an object of class 'scmp'. This object should contain the computed solution.}
}
\value{
Return the terms of the polynomial model.
}
\description{
This function is used to view or return the solution of the provided scMaSigPro object.
}
