% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solShow.R
\name{showSol}
\alias{showSol}
\title{Show Solution Matrix}
\usage{
showSol(scMaSigPro.obj, view = TRUE)
}
\arguments{
\item{scMaSigPro.obj}{A \code{scMaSigPro_t_fit} object containing the results of the ScMaSigPro analysis.}

\item{view}{Logical value indicating whether to view the solution matrix using a data viewer (e.g., RStudio Viewer).}
}
\value{
The solution matrix extracted from the ScMaSigPro analysis results.
}
\description{
This function extracts and displays the solution matrix from the output of the ScMaSigPro
analysis (\code{scMaSigPro_t_fit} object).
}
\keyword{internal}
