% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.sig.genes.class.R
\docType{class}
\name{sigClass}
\alias{sigClass}
\alias{sigClass-class}
\title{Class "sigClass"}
\description{
A class for Storing significant genes.
}
\section{Slots}{

\describe{
\item{\code{summary}}{A dataframe listing significant genes for the variables given
by the function parameters.}

\item{\code{sig.genes}}{A list with detailed information on the significant genes found
for the variables given by the function parameters. Each element of the list is also a list containing:
    \itemize{
      \item{\code{sig.profiles}:}{Expression values of significant genes.}
      \item{\code{coefficients}:}{Regression coefficients of the adjusted models.}
      \item{\code{group.coeffs}:}{Regression coefficients of the implicit models of each experimental group.}
      \item{\code{sig.pvalues}:}{P-values of the regression coefficients for significant genes.}
      \item{\code{g}:}{Number of genes.}
      \item{\code{...}:}{Arguments passed by previous functions.}
    }}
}}

\keyword{classes}
