% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.sig.genes.class.R
\docType{class}
\name{sigClass}
\alias{sigClass}
\alias{sigClass-class}
\title{Class "sigClass"}
\description{
A class for Storing significant genes
}
\section{Slots}{

\describe{
\item{\code{summary}}{A dataframe of genes}

\item{\code{sig.genes}}{A list}
}}

\keyword{classes}
