% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScMaSigProClass.R
\docType{class}
\name{Significant}
\alias{Significant}
\alias{Significant-class}
\title{Significant}
\description{
S4 class to store significantly selected results features and their clusters.
}
\section{Slots}{

\describe{
\item{\code{sig.genes}}{list}

\item{\code{feature.clusters}}{list}
}}

\keyword{classes}
