% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squeeze.R
\name{squeeze}
\alias{squeeze}
\title{Pseudo-bulking with optimal number of pseudotime based bins}
\usage{
squeeze(
  scmpObject,
  pseudotime_colname = scmpObject@addParams@pseudotime_colname,
  path_colname = scmpObject@addParams@path_colname,
  bin_method = "Sturges",
  drop.fac = 0.5,
  verbose = TRUE,
  bin_members_colname = "scmp_bin_members",
  bin_colname = "scmp_bin",
  bin_size_colname = "scmp_bin_size",
  bin_pseudotime_colname = "scmp_binned_pseudotime",
  split_bins = TRUE,
  prune_bins = TRUE,
  assay_name = "counts",
  drop_trails = TRUE,
  cluster_count_by = "sum",
  fill_gaps = TRUE,
  additional_params = list(use_unique_time_points = FALSE)
)
}
\arguments{
\item{scmpObject}{object of Class scMaSigPro. See \code{\link{scMaSigProClass}}
for more details.}

\item{pseudotime_colname}{Name of the column in `cell.metadata` storing
Pseudotime values. Generated using `colData` from the \pkg{SingleCellExperiment}
package. (Default is "Pseudotime").}

\item{path_colname}{Name of the column in `cell.metadata` storing information
for Path. Generated using `colData` from the \pkg{SingleCellExperiment}
package. (Default is `path_prefix`).}

\item{bin_method}{A character string specifying the method to use in order to
estimate the optimal number of bins. Available options: "Freedman.Diaconis",
"Sqrt", "Sturges", "Rice", "Doane", and "Scott.Normal". See \code{\link{estBinSize}}
for more details. (Default = "Sturges").}

\item{drop.fac}{A numeric value specifying the factor by which to decrease the
number of bins if the initial binning results in too many bins. (Default = 0.5).}

\item{verbose}{Print detailed output in the console. (Default is TRUE)
per path iteratively. Options: "universal", "individual. (Default = "universal").}

\item{bin_members_colname}{Name of the column in the 'annotated_cell_metadata'}

\item{bin_colname}{Name of the bin column name}

\item{bin_size_colname}{Setting the name of the bin size column.}

\item{bin_pseudotime_colname}{Name of the column to store the computed Pseudotime
bins.}

\item{split_bins}{If bin sizes are greater than}

\item{prune_bins}{description}

\item{assay_name}{Name of the Assay in the assay_name object from which retrieve the counts.
(Default = "counts").}

\item{drop_trails}{description}

\item{cluster_count_by}{A character string specifying the method to use to
aggregate counts within each cluster. Available options are 'mean' or 'sum'. (Default = "sum").}

\item{fill_gaps}{description}

\item{additional_params}{Pass additional parameters as a named list. See Details.}
}
\value{
A data.frame that contains the original data plus additional columns:
- 'bin' : Name of the bin
- 'bin_size' : Size of the bin
- 'binned_time' : Interval range of each bin
This function returns the merged data.frame with new discretized
pseudotime_colname, preserving the original rownames.
}
\description{
`squeeze()` discretizes a continuous time series column into bins
of equal size using entropy-based binning method. It automatically calculates
the optimal number of bins using one of the supported methods. The bin sizes
are also calculated and merged with the input cell_metadata.
}
\details{
This function performs the following steps:
- Adds a new column 'cell' to the input data.frame which copies the row names.
- Extracts the time series data from the specified column of the input data.frame.
- Calculates the optimal number of bins using the specified method.
- Prints the estimated number of bins if verbose is set to TRUE.
- Discretizes the time series data into bins using the entropy-based binning method.
- Merges the original data.frame with the new binned time series data.
- Removes the 'cell' column and sets the row names back to the original row names of the input data.frame.
- Returns the merged data.frame.
}
\seealso{
\code{\link{estBinSize}}, \code{\link{discretize}}, \code{\link{create_range}}
}
\author{
Priyansh Srivastava \email{spriyansh29@gmail.com}
}
