% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squeeze.R
\name{squeeze}
\alias{squeeze}
\title{The 'squeeze' function}
\usage{
squeeze(
  scmp.ob,
  time.col,
  path.col,
  method = "Sturges",
  drop.fac = 0.5,
  verbose = TRUE,
  cluster.count.by = "sum",
  assay.name = "counts"
)
}
\arguments{
\item{scmp.ob}{An object.}

\item{time.col}{The name of the time column.}

\item{path.col}{The name of the path column.}

\item{method}{Method for data compression (default: "Sturges").}

\item{drop.fac}{The factor by which to drop data (default: 0.5).}

\item{verbose}{Logical. If TRUE, detailed messages are printed (default: TRUE).}

\item{cluster.count.by}{How to count clusters (default: "sum").}

\item{assay.name}{The name of the assay (default: "counts").}
}
\value{
Returns the object 'scmp.ob' with updated slots.
}
\description{
This function performs a series of operations including data compression, creating pseudo-bulk-design, and creating pseudo-bulk counts.
It then generates a SingleCellExperiment object with updated slots.
}
\examples{
# Insert an example of how to use the function here.

}
