% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScMaSigProClass.R
\docType{class}
\name{VariableProfiles}
\alias{VariableProfiles}
\alias{VariableProfiles-class}
\title{VariableProfiles}
\description{
S4 class to store results of the global model fitting for all the features.
}
\section{Slots}{

\describe{
\item{\code{non_flat}}{A character vector of gene names marked as non-flat profiles.}

\item{\code{p_values}}{A numeric vector containing the computed p-values for
non-flat profiles.}

\item{\code{adj_p_values}}{A numeric vector containing the computed adjusted
p-values for non-flat profiles.}

\item{\code{fdr}}{False Discovery Rate (FDR) value for the given significance level.}
}}

\author{
Priyansh Srivastava <spriyansh29@gmail.com>
}
\keyword{classes}
